% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\title{access}
\usage{
access(allow = NULL, deny = NULL, references = NULL, id = NULL,
  system = NULL, scope = NULL, order = NULL, authSystem = NULL)
}
\arguments{
\item{allow}{A rule that grants a permission type.See \code{\link[=allow]{allow()}}}

\item{deny}{A rule that revokes a permission type.See \code{\link[=deny]{deny()}}}

\item{references}{The id of another element in this EML document to be used to here in this context.See \code{\link[=references]{references()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}

\item{order}{The order in which the allow and deny rules should be applied.See \code{\link[=order]{order()}}}

\item{authSystem}{The authentication system is used to verify the user or group to whom access is allowed or denied.See \code{\link[=authSystem]{authSystem()}}}
}
\value{
a access list object
}
\description{
access
}
