% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{accuracy}
\usage{
accuracy(attributeAccuracyReport = NULL,
  quantitativeAttributeAccuracyAssessment = NULL)
}
\arguments{
\item{attributeAccuracyReport}{An explanatory report of the accuracy of the attribute. See \code{\link[=attributeAccuracyReport]{attributeAccuracyReport()}}}

\item{quantitativeAttributeAccuracyAssessment}{A value assigned to summarize the accuracy of the attribute. See \code{\link[=quantitativeAttributeAccuracyAssessment]{quantitativeAttributeAccuracyAssessment()}}}
}
\value{
a accuracy list object
}
\description{
accuracy
}
