% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalMetadata.R
\name{additionalMetadata}
\alias{additionalMetadata}
\title{additionalMetadata}
\usage{
additionalMetadata(describes = NULL, metadata = NULL, id = NULL)
}
\arguments{
\item{describes}{A pointer to the id attribute for the sub-portion of the resource that is described by this additional metadata.}

\item{metadata}{This element contains the additional metadata that is to be included in the document. The content of this field can be any well-formed XML fragment.}

\item{inheritParams}{node_template}
}
\value{
a additionalMetadata list object
}
\description{
additionalMetadata
}
