% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrativeArea.R
\name{administrativeArea}
\alias{administrativeArea}
\title{administrativeArea}
\usage{
administrativeArea(value = NULL, lang = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a administrativeArea list object
}
\description{
administrativeArea
}
