% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allow.R
\name{allow}
\alias{allow}
\title{allow}
\usage{
allow(principal = NULL, permission = NULL)
}
\arguments{
\item{principal}{The user or group (principal) for which the access control applies.}

\item{permission}{The type of permission being granted or denied.}

\item{inheritParams}{node_template}
}
\value{
a allow list object
}
\description{
allow
}
