% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
annotation(propertyURI = NULL, propertyLabel = NULL, valueURI = NULL,
  valueLabel = NULL)
}
\arguments{
\item{propertyURI}{The persistent URI used to identify a property from a vocabulary. See \code{\link[=propertyURI]{propertyURI()}}}

\item{propertyLabel}{A human readable representation of the controlled property. See \code{\link[=propertyLabel]{propertyLabel()}}}

\item{valueURI}{The persistent URI used to identify a value from a vocabulary. See \code{\link[=valueURI]{valueURI()}}}

\item{valueLabel}{A human readable representation of the controlled value. See \code{\link[=valueLabel]{valueLabel()}}}
}
\value{
a annotation list object
}
\description{
annotation
}
