% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{article}
\alias{article}
\title{article}
\usage{
article(journal = NULL, volume = NULL, issue = NULL,
  pageRange = NULL, publisher = NULL, publicationPlace = NULL,
  ISSN = NULL)
}
\arguments{
\item{journal}{The name of the journal, magazine, newspaper, zine, etc... in which the article was published.}

\item{volume}{The volume of the journal in which the article appears}

\item{issue}{The issue of the journal in which the article appears}

\item{pageRange}{The beginning and end page numbers of the journal article}

\item{publisher}{Organization that actually publishes the article}

\item{publicationPlace}{The location at which the work was published.}

\item{ISSN}{The unique Internation Standard Serial Number}

\item{inheritParams}{node_template}
}
\value{
a article list object
}
\description{
article
}
