% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeList.R
\name{attributeList}
\alias{attributeList}
\title{attributeList}
\usage{
attributeList(attribute = NULL, references = NULL, id = NULL)
}
\arguments{
\item{attribute}{(def not found)}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a attributeList list object
}
\description{
attributeList
}
