% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authentication}
\alias{authentication}
\title{authentication}
\usage{
authentication(method = NULL)
}
\arguments{
\item{method}{The method used to calculate an authentication checksum.See \code{\link[=method]{method()}}}
}
\value{
a authentication list object
}
\description{
authentication
}
