% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/award.R
\name{award}
\alias{award}
\title{award}
\usage{
award(funderName = NULL, funderIdentifier = NULL, awardNumber = NULL,
  title = NULL, awardUrl = NULL)
}
\arguments{
\item{funderName}{The name of the funding institution that made this awardSee \code{\link[=funderName]{funderName()}}}

\item{funderIdentifier}{The identifier of the funding agency.See \code{\link[=funderIdentifier]{funderIdentifier()}}}

\item{awardNumber}{The assigned award number.See \code{\link[=awardNumber]{awardNumber()}}}

\item{title}{The title of the award.See \code{\link[=title]{title()}}}

\item{awardUrl}{The URL associated with award.See \code{\link[=awardUrl]{awardUrl()}}}
}
\value{
a award list object
}
\description{
award
}
