% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awardNumber.R
\name{awardNumber}
\alias{awardNumber}
\title{awardNumber}
\usage{
awardNumber(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a awardNumber list object
}
\description{
awardNumber
}
