% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beginDate.R
\name{beginDate}
\alias{beginDate}
\title{beginDate}
\usage{
beginDate(calendarDate = NULL, time = NULL,
  alternativeTimeScale = NULL)
}
\arguments{
\item{calendarDate}{The calendar date for an event.}

\item{time}{The time of day for an event.}

\item{alternativeTimeScale}{A name, code, or date describing an event or period in an alternative time scale, such as one of the geologic time scales.}

\item{inheritParams}{node_template}
}
\value{
a beginDate list object
}
\description{
beginDate
}
