% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryRasterFormat.R
\name{binaryRasterFormat}
\alias{binaryRasterFormat}
\title{binaryRasterFormat}
\usage{
binaryRasterFormat(rowColumnOrientation = NULL, multiBand = NULL,
  nbits = NULL, byteorder = NULL, skipbytes = NULL,
  bandrowbytes = NULL, totalrowbytes = NULL, bandgapbytes = NULL)
}
\arguments{
\item{rowColumnOrientation}{Orientation for reading rows and columns.}

\item{multiBand}{Multiple band image information.}

\item{nbits}{The number of bits per pixel per band.}

\item{byteorder}{The byte order in which values are stored.}

\item{skipbytes}{The number of bytes of data in the image file to skip in order to reach the start of the image data.}

\item{bandrowbytes}{The number of bytes per band per row.}

\item{totalrowbytes}{The total number of bytes of data per row.}

\item{bandgapbytes}{The number of bytes between bands in a BSQ format image.}

\item{inheritParams}{node_template}
}
\value{
a binaryRasterFormat list object
}
\description{
binaryRasterFormat
}
