% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundingAltitudes.R
\name{boundingAltitudes}
\alias{boundingAltitudes}
\title{boundingAltitudes}
\usage{
boundingAltitudes(altitudeMinimum = NULL, altitudeMaximum = NULL,
  altitudeUnits = NULL)
}
\arguments{
\item{altitudeMinimum}{The minimum altitude extent of coverage.}

\item{altitudeMaximum}{The maximum altitude extent of coverage.}

\item{altitudeUnits}{The unit of altitude}

\item{inheritParams}{node_template}
}
\value{
a boundingAltitudes list object
}
\description{
boundingAltitudes
}
