% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundingCoordinates.R
\name{boundingCoordinates}
\alias{boundingCoordinates}
\title{boundingCoordinates}
\usage{
boundingCoordinates(westBoundingCoordinate = NULL,
  eastBoundingCoordinate = NULL, northBoundingCoordinate = NULL,
  southBoundingCoordinate = NULL, boundingAltitudes = NULL)
}
\arguments{
\item{westBoundingCoordinate}{Western-most limit of a bounding box, expressed in degrees of longitude.}

\item{eastBoundingCoordinate}{Eastern-most limit of a bounding box, expressed in degrees of longitude.}

\item{northBoundingCoordinate}{Northern-most lilmit of a bounding box expressed in latitude.}

\item{southBoundingCoordinate}{Southern-most limit of the bounding box expressed in latitude.}

\item{boundingAltitudes}{The vertical limits of a data set expressed by altitude.}

\item{inheritParams}{node_template}
}
\value{
a boundingCoordinates list object
}
\description{
boundingCoordinates
}
