% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\title{bounds}
\usage{
bounds(minimum = NULL, maximum = NULL)
}
\arguments{
\item{minimum}{Minimum numeric bound of attribute}

\item{maximum}{Maximum numeric bound of attribute}

\item{inheritParams}{node_template}

\item{minimum}{Minimum date bound of attribute}

\item{maximum}{Maximum date bound of attribute}
}
\value{
a bounds list object
}
\description{
bounds
}
