% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeHistory.R
\name{changeHistory}
\alias{changeHistory}
\title{changeHistory}
\usage{
changeHistory(changeScope = NULL, oldValue = NULL, changeDate = NULL,
  comment = NULL)
}
\arguments{
\item{changeScope}{An expression describing the scope to which the documented change was applied.}

\item{oldValue}{The previous value of the data prior to the change.}

\item{changeDate}{The date the changes were applied.}

\item{comment}{Explanation or justification for the change made to the data.}

\item{inheritParams}{node_template}
}
\value{
a changeHistory list object
}
\description{
changeHistory
}
