% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConstraint.R
\name{checkConstraint}
\alias{checkConstraint}
\title{checkConstraint}
\usage{
checkConstraint(constraintName = NULL, constraintDescription = NULL,
  checkCondition = NULL, language = NULL)
}
\arguments{
\item{constraintName}{A meaningfull name of the constraint.}

\item{constraintDescription}{Descibes the purpose of the constraint.}

\item{checkCondition}{An SQL statement or other language implementation of the condition for a check constraint.}

\item{language}{The language that the is used to express or implement the check constraint.}

\item{inheritParams}{node_template}
}
\value{
a checkConstraint list object
}
\description{
checkConstraint
}
