% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citetitle.R
\name{citetitle}
\alias{citetitle}
\title{citetitle}
\usage{
citetitle(lang = NULL)
}
\arguments{
\item{lang}{language code for element contentSee \code{\link[=lang]{lang()}}}
}
\value{
a citetitle list object
}
\description{
citetitle
}
