% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{city}
\alias{city}
\title{city}
\usage{
city(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a city list object
}
\description{
city
}
