% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connection}
\alias{connection}
\title{connection}
\usage{
connection(connectionDefinition = NULL, parameter = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{connectionDefinition}{Connection Definition}

\item{parameter}{A parameter to be used to make this connection.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a connection list object
}
\description{
connection
}
