% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.R
\name{constraint}
\alias{constraint}
\title{constraint}
\usage{
constraint(primaryKey = NULL, uniqueKey = NULL,
  checkConstraint = NULL, foreignKey = NULL, joinCondition = NULL,
  notNullConstraint = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{primaryKey}{The primary key in the entity}

\item{uniqueKey}{A unique key in the entity}

\item{checkConstraint}{A constraint which checks a conditional clause within an entity.}

\item{foreignKey}{A foreign key relationship among entities}

\item{joinCondition}{A non primary/foreign key join}

\item{notNullConstraint}{A constraint that indicates that no null values should be present for an attribute.}

\item{inheritParams}{node_template}
}
\value{
a constraint list object
}
\description{
constraint
}
