% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlPoint.R
\name{controlPoint}
\alias{controlPoint}
\title{controlPoint}
\usage{
controlPoint(column = NULL, row = NULL, xCoordinate = NULL,
  yCoordinate = NULL, pointInPixel = NULL)
}
\arguments{
\item{column}{Column location of the georeferencing point in the grid.}

\item{row}{Row location of the georeferencing point in the grid.}

\item{xCoordinate}{X Location of the georeferencing point in the X diminsion of the coordinate system.}

\item{yCoordinate}{Location of the georeferencing point in the Y dimension of the coordinate system.}

\item{pointInPixel}{Location within the pixel of the georeferencing point.}

\item{inheritParams}{node_template}
}
\value{
a controlPoint list object
}
\description{
controlPoint
}
