% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornerPoint.R
\name{cornerPoint}
\alias{cornerPoint}
\title{cornerPoint}
\usage{
cornerPoint(xCoordinate = NULL, yCoordinate = NULL,
  pointInPixel = NULL, corner = NULL)
}
\arguments{
\item{xCoordinate}{X Location of the georeferencing point in the X diminsion of the coordinate system.See \code{\link[=xCoordinate]{xCoordinate()}}}

\item{yCoordinate}{Location of the georeferencing point in the Y dimension of the coordinate system.See \code{\link[=yCoordinate]{yCoordinate()}}}

\item{pointInPixel}{Location within the pixel of the georeferencing point.See \code{\link[=pointInPixel]{pointInPixel()}}}

\item{corner}{Identification of the corner in the grid corresponding to the coordinates provided.See \code{\link[=corner]{corner()}}}
}
\value{
a cornerPoint list object
}
\description{
cornerPoint
}
