% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{coverage}
\usage{
coverage(geographicCoverage = NULL, temporalCoverage = NULL,
  taxonomicCoverage = NULL, references = NULL, id = NULL,
  system = NULL, scope = NULL, geographicDescription = NULL,
  boundingCoordinates = NULL, datasetGPolygon = NULL)
}
\arguments{
\item{geographicCoverage}{Geographic coverage information.See \code{\link[=geographicCoverage]{geographicCoverage()}}}

\item{temporalCoverage}{Temporal coverage information.See \code{\link[=temporalCoverage]{temporalCoverage()}}}

\item{taxonomicCoverage}{Taxonomic coverage information.See \code{\link[=taxonomicCoverage]{taxonomicCoverage()}}}

\item{references}{The id of another element in this EML document to be used to here in this context.See \code{\link[=references]{references()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}

\item{geographicDescription}{Short text description of the geographic areal domain of the data set.See \code{\link[=geographicDescription]{geographicDescription()}}}

\item{boundingCoordinates}{The four margins (N, S, E, W) of a bounding box on the earth's surface, or when considered in lat-lon pairs, the corners of the box. To define a single point, use the same value in each lat or lon pair. These elements are meant to convey general information and are not for accurate mapping. More specific information may be included by using the elements in the spatialReference schema.See \code{\link[=boundingCoordinates]{boundingCoordinates()}}}

\item{datasetGPolygon}{This construct creates a spatial ring with a hollow center.See \code{\link[=datasetGPolygon]{datasetGPolygon()}}}
}
\value{
a coverage list object
}
\description{
coverage
}
