% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creator.R
\name{creator}
\alias{creator}
\title{creator}
\usage{
creator(individualName = NULL, organizationName = NULL,
  positionName = NULL, address = NULL, phone = NULL,
  electronicMailAddress = NULL, onlineUrl = NULL, userId = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{individualName}{The full name of the person being described}

\item{organizationName}{The full name of the organization being described}

\item{positionName}{The name of the title or position associated with the resource.}

\item{address}{The full address information for a given responsible party entry.}

\item{phone}{Information about the contact's telephone}

\item{electronicMailAddress}{The email address of the contact.}

\item{onlineUrl}{A link to associated online information, usually a web site.}

\item{userId}{An identifier that links this party to a directory of individuals}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a creator list object
}
\description{
creator
}
