% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{dataset}
\usage{
dataset(alternateIdentifier = NULL, shortName = NULL, title = NULL,
  creator = NULL, metadataProvider = NULL, associatedParty = NULL,
  pubDate = NULL, language = NULL, series = NULL, abstract = NULL,
  keywordSet = NULL, additionalInfo = NULL,
  intellectualRights = NULL, licensed = NULL, distribution = NULL,
  coverage = NULL, annotation = NULL, purpose = NULL,
  introduction = NULL, gettingStarted = NULL,
  acknowledgements = NULL, maintenance = NULL, contact = NULL,
  publisher = NULL, pubPlace = NULL, methods = NULL,
  project = NULL, dataTable = NULL, spatialRaster = NULL,
  spatialVector = NULL, storedProcedure = NULL, view = NULL,
  otherEntity = NULL, referencePublication = NULL,
  usageCitation = NULL, literatureCited = NULL, references = NULL,
  id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity}

\item{shortName}{A short name that describes the resource, sometimes a filename.}

\item{title}{A brief description of the resource, providing enough detail to differentiate it from other similar resources.}

\item{creator}{The people or organizations who created this resource.}

\item{metadataProvider}{The people or organizations who created provided documentation and other metadata for this resource.}

\item{associatedParty}{Other people or organizations who should be associated with this resource.}

\item{pubDate}{The publication date of the resource.}

\item{language}{The language in which the resource is written.}

\item{series}{The series from which the resource came.}

\item{abstract}{A brief overview of the resource.}

\item{keywordSet}{Keyword information that describes the resource.}

\item{additionalInfo}{Any extra information pertitent to the resource.}

\item{intellectualRights}{Intellectual property rights regarding usage and licensing of this resource.}

\item{licensed}{Information identifying a well-known license for the metadata and data}

\item{distribution}{Information on how the resource is distributed online and offline}

\item{coverage}{Extent of the coverage of the resource.}

\item{annotation}{A precisely-defined semantic statement about this resource.}

\item{purpose}{A synopsis of the purpose of this dataset.}

\item{introduction}{An overview of the background and context for the dataset.}

\item{gettingStarted}{A high level overview of interpretation, structure, and content of the dataset.}

\item{acknowledgements}{Text that acknowledges funders and other key contributors.}

\item{maintenance}{A description of the maintenance of this data resource.}

\item{contact}{The contact for this dataset}

\item{publisher}{The publisher of this data set}

\item{pubPlace}{The location that the resource was published.}

\item{methods}{The methods field documents scientific methods used in the collection of this dataset.}

\item{project}{The project field contains information on the project in which this dataset was collected.}

\item{dataTable}{The dataTable field documents the dataTable(s) that make up this dataset.}

\item{spatialRaster}{The spatialRaster field describes any spatial raster images included in this dataset.}

\item{spatialVector}{The spatialVector field describes any spatial vectors included in this dataset.}

\item{storedProcedure}{The storedProcedure field contains information about any stored procedures included with this dataset.}

\item{view}{The view field contains information about any views included with this dataset.}

\item{otherEntity}{The otherEntity field contains information about any entity in the dataset that is not any of the preceding entities.}

\item{referencePublication}{A citation to an additional publication that serves as an important reference for a dataset.}

\item{usageCitation}{A citation to articles or products in which the dataset is used or referenced.}

\item{literatureCited}{A citation to articles or products which were referenced in the dataset or its associated metadata.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a dataset list object
}
\description{
dataset
}
