% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetGPolygon.R
\name{datasetGPolygon}
\alias{datasetGPolygon}
\title{datasetGPolygon}
\usage{
datasetGPolygon(datasetGPolygonOuterGRing = NULL,
  datasetGPolygonExclusionGRing = NULL)
}
\arguments{
\item{datasetGPolygonOuterGRing}{The outer containment loop of a datasetGPolygon.See \code{\link[=datasetGPolygonOuterGRing]{datasetGPolygonOuterGRing()}}}

\item{datasetGPolygonExclusionGRing}{Data Set G-Polygon Exclusion G-Ring, the closed nonintersecting boundary of a void area (or hole in an interior area).See \code{\link[=datasetGPolygonExclusionGRing]{datasetGPolygonExclusionGRing()}}}
}
\value{
a datasetGPolygon list object
}
\description{
datasetGPolygon
}
