% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetGPolygonExclusionGRing.R
\name{datasetGPolygonExclusionGRing}
\alias{datasetGPolygonExclusionGRing}
\title{datasetGPolygonExclusionGRing}
\usage{
datasetGPolygonExclusionGRing(gRingPoint = NULL, gRing = NULL)
}
\arguments{
\item{gRingPoint}{A single geographic locationSee \code{\link[=gRingPoint]{gRingPoint()}}}
}
\value{
a datasetGPolygonExclusionGRing list object
}
\description{
datasetGPolygonExclusionGRing
}
