% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetGPolygonOuterGRing.R
\name{datasetGPolygonOuterGRing}
\alias{datasetGPolygonOuterGRing}
\title{datasetGPolygonOuterGRing}
\usage{
datasetGPolygonOuterGRing(gRingPoint = NULL, gRing = NULL)
}
\arguments{
\item{gRingPoint}{A single geographic location}

\item{gRing}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a datasetGPolygonOuterGRing list object
}
\description{
datasetGPolygonOuterGRing
}
