% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTime.R
\name{dateTime}
\alias{dateTime}
\title{dateTime}
\usage{
dateTime(formatString = NULL, dateTimePrecision = NULL,
  dateTimeDomain = NULL)
}
\arguments{
\item{formatString}{A format string that describes the format for a date-time value from the Gregorian calendar. See \code{\link[=formatString]{formatString()}}}

\item{dateTimePrecision}{An indication of the precision of a date or time value See \code{\link[=dateTimePrecision]{dateTimePrecision()}}}

\item{dateTimeDomain}{See the summary for the type: DateTimeDomainType See \code{\link[=dateTimeDomain]{dateTimeDomain()}}}
}
\value{
a dateTime list object
}
\description{
dateTime
}
