% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryPoint.R
\name{deliveryPoint}
\alias{deliveryPoint}
\title{deliveryPoint}
\usage{
deliveryPoint(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a deliveryPoint list object
}
\description{
deliveryPoint
}
