% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency.R
\name{dependency}
\alias{dependency}
\title{dependency}
\usage{
dependency(action = NULL, software = NULL)
}
\arguments{
\item{action}{Describes what action needs to be undertaken (if any) for a software dependency at either the software package or implementation level.}

\item{software}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a dependency list object
}
\description{
dependency
}
