% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthSysDef.R
\name{depthSysDef}
\alias{depthSysDef}
\title{depthSysDef}
\usage{
depthSysDef(depthDatumName = NULL, depthResolution = NULL,
  depthDistanceUnits = NULL, depthEncodingMethod = NULL)
}
\arguments{
\item{depthDatumName}{The identification given to surface of reference from which depths are measured.}

\item{depthResolution}{The minimum distance possible between two adjacent depth values, expressed in Depth Distance Units of measure.}

\item{depthDistanceUnits}{The units in which depths are recorded.}

\item{depthEncodingMethod}{The means used to encode depths.}

\item{inheritParams}{node_template}
}
\value{
a depthSysDef list object
}
\description{
depthSysDef
}
