% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor.R
\name{descriptor}
\alias{descriptor}
\title{descriptor}
\usage{
descriptor(descriptorValue = NULL, citation = NULL, name = NULL,
  citableClassificationSystem = NULL)
}
\arguments{
\item{descriptorValue}{Description of some aspect of the study area.}

\item{citation}{A citation for this descriptor.}

\item{name}{The name of the descriptor system.}

\item{citableClassificationSystem}{This boolean attribute defines whether this descriptor comes from a citable classification system or not.}

\item{inheritParams}{node_template}
}
\value{
a descriptor list object
}
\description{
descriptor
}
