% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptorValue.R
\name{descriptorValue}
\alias{descriptorValue}
\title{descriptorValue}
\usage{
descriptorValue(name_or_id = NULL)
}
\arguments{
\item{name_or_id}{The name part of a name/value pair of a descriptor; or ID portion of a classification, if applicable.}
}
\value{
a descriptorValue list object
}
\description{
descriptorValue
}
