% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designDescription.R
\name{designDescription}
\alias{designDescription}
\title{designDescription}
\usage{
designDescription(description = NULL, citation = NULL)
}
\arguments{
\item{description}{Textual description of research design. See \code{\link[=description]{description()}}}

\item{citation}{Citation that describes the research design. See \code{\link[=citation]{citation()}}}
}
\value{
a designDescription list object
}
\description{
designDescription
}
