% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{distribution}
\usage{
distribution(online = NULL, offline = NULL, inline = NULL,
  access = NULL, references = NULL, id = NULL, system = NULL,
  scope = NULL)
}
\arguments{
\item{online}{online}

\item{offline}{offline}

\item{inline}{inline}

\item{access}{access}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}

\item{online}{Online distribution information}

\item{offline}{data are available offline}

\item{inline}{data distributed inline in the metadata.}
}
\value{
a distribution list object
}
\description{
distribution
}
