% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editedBook.R
\name{editedBook}
\alias{editedBook}
\title{editedBook}
\usage{
editedBook(publisher = NULL, publicationPlace = NULL, edition = NULL,
  volume = NULL, numberOfVolumes = NULL, totalPages = NULL,
  totalFigures = NULL, totalTables = NULL, ISBN = NULL)
}
\arguments{
\item{publisher}{Organization that actually publishes the book}

\item{publicationPlace}{The location at which the work was published.}

\item{edition}{The edition of the book being described.}

\item{volume}{The volume of the book that is part of a series.}

\item{numberOfVolumes}{Number of volumes in a collection}

\item{totalPages}{The total number of pages in the book.}

\item{totalFigures}{The total number of figures in the book.}

\item{totalTables}{The total number of tables in a book.}

\item{ISBN}{The unique Internation Standard Book Number}

\item{inheritParams}{node_template}
}
\value{
a editedBook list object
}
\description{
editedBook
}
