% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml.R
\name{eml}
\alias{eml}
\title{eml}
\usage{
eml(access = NULL, dataset = NULL, citation = NULL,
  software = NULL, protocol = NULL, additionalMetadata = NULL,
  packageId = NULL, system = NULL, scope = NULL, lang = NULL)
}
\arguments{
\item{access}{Access control rules for the entire resource, which can be overridden by access rules in distribution trees}

\item{dataset}{A resource that describes a data set, which can include one or more data entities such as data tables.}

\item{citation}{A resource that describes a literature citation that one might find in a bibliography.}

\item{software}{A resource that describes a software package, which can include commercial and non-commercial software as well as data processing programs.}

\item{protocol}{A resource that describes a scientific protocol, which can include one or more descriptions of methods and procedures.}

\item{additionalMetadata}{A flexible field for including any other relevant metadata that pertains to the resource being described.}

\item{packageId}{A unique identifier for this entire EML metadata document that can be used to reference it elsewhere.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a eml list object
}
\description{
eml
}
