% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphasis.R
\name{emphasis}
\alias{emphasis}
\title{emphasis}
\usage{
emphasis(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a emphasis list object
}
\description{
emphasis
}
