% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeratedDomain.R
\name{enumeratedDomain}
\alias{enumeratedDomain}
\title{enumeratedDomain}
\usage{
enumeratedDomain(codeDefinition = NULL, externalCodeSet = NULL,
  entityCodeList = NULL, enforced = NULL)
}
\arguments{
\item{codeDefinition}{A codes and its definition}

\item{externalCodeSet}{A reference to an externally defined set of codes used in this attribute}

\item{entityCodeList}{A code list that is defined in a data table}

\item{enforced}{Indicates whether the enumerated domain values enforced.}

\item{inheritParams}{node_template}
}
\value{
a enumeratedDomain list object
}
\description{
enumeratedDomain
}
