% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externallyDefinedFormat.R
\name{externallyDefinedFormat}
\alias{externallyDefinedFormat}
\title{externallyDefinedFormat}
\usage{
externallyDefinedFormat(formatName = NULL, formatVersion = NULL,
  citation = NULL)
}
\arguments{
\item{formatName}{Name of the format of the data object}

\item{formatVersion}{Version of the format of the data object}

\item{citation}{Citation providing more details about the physical format.}

\item{inheritParams}{node_template}
}
\value{
a externallyDefinedFormat list object
}
\description{
externallyDefinedFormat
}
