% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreignKey.R
\name{foreignKey}
\alias{foreignKey}
\title{foreignKey}
\usage{
foreignKey(key = NULL, entityReference = NULL,
  relationshipType = NULL, cardinality = NULL)
}
\arguments{
\item{key}{The set of attributes to which this constraint applies.}

\item{entityReference}{The id of the parent-entity in a foreign key constraint.}

\item{relationshipType}{Relationship type: Identifying or non-identifying}

\item{cardinality}{Cardinality of the relationship between a parent entity and a child entity.}

\item{inheritParams}{node_template}
}
\value{
a foreignKey list object
}
\description{
foreignKey
}
