% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funderIdentifier.R
\name{funderIdentifier}
\alias{funderIdentifier}
\title{funderIdentifier}
\usage{
funderIdentifier(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a funderIdentifier list object
}
\description{
funderIdentifier
}
