% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{funding}
\alias{funding}
\title{funding}
\usage{
funding(section = NULL, para = NULL, markdown = NULL, lang = NULL)
}
\arguments{
\item{section}{A section of related text.}

\item{para}{A simple paragraph of text.}

\item{markdown}{A block of text formatted with Markdown directives.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a funding list object
}
\description{
funding
}
