% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicCoverage.R
\name{geographicCoverage}
\alias{geographicCoverage}
\title{geographicCoverage}
\usage{
geographicCoverage(geographicDescription = NULL,
  boundingCoordinates = NULL, datasetGPolygon = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{geographicDescription}{Short text description of the geographic areal domain of the data set.}

\item{boundingCoordinates}{The four margins (N, S, E, W) of a bounding box on the earth's surface, or when considered in lat-lon pairs, the corners of the box. To define a single point, use the same value in each lat or lon pair. These elements are meant to convey general information and are not for accurate mapping. More specific information may be included by using the elements in the spatialReference schema.}

\item{datasetGPolygon}{This construct creates a spatial ring with a hollow center.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a geographicCoverage list object
}
\description{
geographicCoverage
}
