% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeferenceInfo.R
\name{georeferenceInfo}
\alias{georeferenceInfo}
\title{georeferenceInfo}
\usage{
georeferenceInfo(cornerPoint = NULL, controlPoint = NULL,
  bilinearFit = NULL)
}
\arguments{
\item{cornerPoint}{Location of a corner on a rectified grid within the coordinate system. See \code{\link[=cornerPoint]{cornerPoint()}}}

\item{controlPoint}{A point defined in both grid cell and georeferenced coordinates that is used to position the raster grid within the coordinate system defined in the spatialReference modoule. See \code{\link[=controlPoint]{controlPoint()}}}

\item{bilinearFit}{Terms of a bilinear model to fit the grid to the coordinate system. See \code{\link[=bilinearFit]{bilinearFit()}}}
}
\value{
a georeferenceInfo list object
}
\description{
georeferenceInfo
}
