% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettingStarted.R
\name{gettingStarted}
\alias{gettingStarted}
\title{gettingStarted}
\usage{
gettingStarted(section = NULL, para = NULL, markdown = NULL,
  lang = NULL)
}
\arguments{
\item{section}{A section of related text.}

\item{para}{A simple paragraph of text.}

\item{markdown}{A block of text formatted with Markdown directives.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a gettingStarted list object
}
\description{
gettingStarted
}
