% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givenName.R
\name{givenName}
\alias{givenName}
\title{givenName}
\usage{
givenName(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a givenName list object
}
\description{
givenName
}
