% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizCoordSysDef.R
\name{horizCoordSysDef}
\alias{horizCoordSysDef}
\title{horizCoordSysDef}
\usage{
horizCoordSysDef(geogCoordSys = NULL, projCoordSys = NULL,
  name = NULL)
}
\arguments{
\item{geogCoordSys}{The coodinate system used to express the geographic coordinates. See \code{\link[=geogCoordSys]{geogCoordSys()}}}

\item{projCoordSys}{The coodinate system used to express planar coordinates. See \code{\link[=projCoordSys]{projCoordSys()}}}
}
\value{
a horizCoordSysDef list object
}
\description{
horizCoordSysDef
}
