% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontalAccuracy.R
\name{horizontalAccuracy}
\alias{horizontalAccuracy}
\title{horizontalAccuracy}
\usage{
horizontalAccuracy(accuracyReport = NULL,
  quantitativeAccuracyReport = NULL)
}
\arguments{
\item{accuracyReport}{A qualitative statement about the accuracy of the data. See \code{\link[=accuracyReport]{accuracyReport()}}}

\item{quantitativeAccuracyReport}{A quantitative assessment of the data quality. See \code{\link[=quantitativeAccuracyReport]{quantitativeAccuracyReport()}}}
}
\value{
a horizontalAccuracy list object
}
\description{
horizontalAccuracy
}
