% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageDescription.R
\name{imageDescription}
\alias{imageDescription}
\title{imageDescription}
\usage{
imageDescription(illuminationElevationAngle = NULL,
  illuminationAzimuthAngle = NULL, imageOrientationAngle = NULL,
  imagingCondition = NULL, imageQualityCode = NULL,
  cloudCoverPercentage = NULL, preProcessingTypeCode = NULL,
  compressionGenerationQuality = NULL, triangulationIndicator = NULL,
  radiometricDataAvailability = NULL,
  cameraCalibrationInformationAvailability = NULL,
  filmDistortionInformationAvailability = NULL,
  lensDistortionInformationAvailability = NULL, bandDescription = NULL)
}
\arguments{
\item{illuminationElevationAngle}{Illumination elevation measured in degrees clockwise from the target plane at intersection of the optical line of sight with the earth's surface.}

\item{illuminationAzimuthAngle}{Illumination azimuth measured in degrees clockwise from true north at the time the image is taken.}

\item{imageOrientationAngle}{Angle from the first row of the image to true north in degrees, clockwise.}

\item{imagingCondition}{Code which indicates conditions which affect the quality of the image.}

\item{imageQualityCode}{Specifies the image quality.}

\item{cloudCoverPercentage}{Area of the dataset obscured by clouds, expressed as a percentage of the spatial extent.}

\item{preProcessingTypeCode}{Image distributor's code that identifies the level of radiometric and geometric processing applied against the image.}

\item{compressionGenerationQuality}{Counts the number of lossy compression cycles performed on the image.}

\item{triangulationIndicator}{Code which indicates whether or not triangulation has been performed upon the image.}

\item{radiometricDataAvailability}{Code which indicates whether or not Standard Radiometric Product data is available.}

\item{cameraCalibrationInformationAvailability}{Code which indicates whether or not constants are available which allow for camera calibration corrections.}

\item{filmDistortionInformationAvailability}{Code which indicates whether or not Calibration Reseau information is available}

\item{lensDistortionInformationAvailability}{Code which indicates whether or not lens aberration correction information is available}

\item{bandDescription}{Description of the wavelength domain that the sensor operates in.}

\item{inheritParams}{node_template}
}
\value{
a imageDescription list object
}
\description{
imageDescription
}
