% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individualName.R
\name{individualName}
\alias{individualName}
\title{individualName}
\usage{
individualName(salutation = NULL, givenName = NULL, surName = NULL)
}
\arguments{
\item{salutation}{The salutation used to address an individual}

\item{givenName}{The given name of the individual.}

\item{surName}{The last name of the individual.}

\item{inheritParams}{node_template}
}
\value{
a individualName list object
}
\description{
individualName
}
