% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intellectualRights.R
\name{intellectualRights}
\alias{intellectualRights}
\title{intellectualRights}
\usage{
intellectualRights(section = NULL, para = NULL, markdown = NULL,
  lang = NULL)
}
\arguments{
\item{section}{A section of related text. See \code{\link[=section]{section()}}}

\item{para}{A simple paragraph of text. See \code{\link[=para]{para()}}}

\item{markdown}{A block of text formatted with Markdown directives. See \code{\link[=markdown]{markdown()}}}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a intellectualRights list object
}
\description{
intellectualRights
}
