% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval.R
\name{interval}
\alias{interval}
\title{interval}
\usage{
interval(unit = NULL, precision = NULL, numericDomain = NULL)
}
\arguments{
\item{unit}{(def not found)}

\item{precision}{(def not found)}

\item{numericDomain}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a interval list object
}
\description{
interval
}
