% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemizedlist.R
\name{itemizedlist}
\alias{itemizedlist}
\title{itemizedlist}
\usage{
itemizedlist(listitem = NULL)
}
\arguments{
\item{listitem}{An item in a list of items.}

\item{inheritParams}{node_template}
}
\value{
a itemizedlist list object
}
\description{
itemizedlist
}
