% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinCondition.R
\name{joinCondition}
\alias{joinCondition}
\title{joinCondition}
\usage{
joinCondition(key = NULL, entityReference = NULL,
  relationshipType = NULL, cardinality = NULL, referencedKey = NULL)
}
\arguments{
\item{key}{The set of attributes to which this constraint applies.}

\item{entityReference}{The id of the parent-entity in a foreign key constraint.}

\item{relationshipType}{Relationship type: Identifying or non-identifying}

\item{cardinality}{Cardinality of the relationship between a parent entity and a child entity.}

\item{referencedKey}{The set of attributes to which a foreign key constraint refers.}

\item{inheritParams}{node_template}
}
\value{
a joinCondition list object
}
\description{
joinCondition
}
