% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\title{key}
\usage{
key(attributeReference = NULL)
}
\arguments{
\item{attributeReference}{The identifier of an attribute found in the identified entity}

\item{inheritParams}{node_template}
}
\value{
a key list object
}
\description{
key
}
