% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{keyword}
\alias{keyword}
\title{keyword}
\usage{
keyword(value = NULL, lang = NULL, keywordType = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{keywordType}{The type of each keyword.}

\item{inheritParams}{node_template}
}
\value{
a keyword list object
}
\description{
keyword
}
