% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language}
\alias{language}
\title{language}
\usage{
language(value = NULL, lang = NULL, LanguageValue = NULL,
  LanguageCodeStandard = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{LanguageValue}{The actual value for the language or a code for the language.}

\item{LanguageCodeStandard}{The International Language Code being used in the field languageValue.}

\item{inheritParams}{node_template}
}
\value{
a language list object
}
\description{
language
}
