% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensed.R
\name{licensed}
\alias{licensed}
\title{licensed}
\usage{
licensed(licenseName = NULL, url = NULL, identifier = NULL)
}
\arguments{
\item{licenseName}{The official name of the license}

\item{url}{The persistent URL for the license}

\item{identifier}{License Identifier}

\item{inheritParams}{node_template}
}
\value{
a licensed list object
}
\description{
licensed
}
