% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listitem.R
\name{listitem}
\alias{listitem}
\title{listitem}
\usage{
listitem(para = NULL, itemizedlist = NULL, orderedlist = NULL)
}
\arguments{
\item{para}{A simple paragraph of text.See \code{\link[=para]{para()}}}

\item{itemizedlist}{A list of items in a text paragraph.See \code{\link[=itemizedlist]{itemizedlist()}}}

\item{orderedlist}{An ordered list of items in a text paragraph.See \code{\link[=orderedlist]{orderedlist()}}}
}
\value{
a listitem list object
}
\description{
listitem
}
