% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literalLayout.R
\name{literalLayout}
\alias{literalLayout}
\title{literalLayout}
\usage{
literalLayout(value = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}
}
\value{
a literalLayout list object
}
\description{
literalLayout
}
