% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literatureCited.R
\name{literatureCited}
\alias{literatureCited}
\title{literatureCited}
\usage{
literatureCited(citation = NULL, bibtex = NULL)
}
\arguments{
\item{citation}{A citation within a list of citations.}

\item{bibtex}{List of citations in Bibtex format.}

\item{inheritParams}{node_template}
}
\value{
a literatureCited list object
}
\description{
literatureCited
}
