% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance.R
\name{maintenance}
\alias{maintenance}
\title{maintenance}
\usage{
maintenance(description = NULL, maintenanceUpdateFrequency = NULL,
  changeHistory = NULL)
}
\arguments{
\item{description}{A text description of the maintenance of this data resource.See \code{\link[=description]{description()}}}

\item{maintenanceUpdateFrequency}{Frequency with which changes and additions are made to the dataset after the initial dataset is completed.See \code{\link[=maintenanceUpdateFrequency]{maintenanceUpdateFrequency()}}}

\item{changeHistory}{A description of changes made to the data since its release.See \code{\link[=changeHistory]{changeHistory()}}}
}
\value{
a maintenance list object
}
\description{
maintenance
}
