% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript.R
\name{manuscript}
\alias{manuscript}
\title{manuscript}
\usage{
manuscript(institution = NULL, totalPages = NULL)
}
\arguments{
\item{institution}{The address and contact information needed to request a manuscript.See \code{\link[=institution]{institution()}}}

\item{totalPages}{The total number of pages in the manuscript.See \code{\link[=totalPages]{totalPages()}}}
}
\value{
a manuscript list object
}
\description{
manuscript
}
