% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(publisher = NULL, edition = NULL, geographicCoverage = NULL,
  scale = NULL)
}
\arguments{
\item{publisher}{Organization that actually publishes the map}

\item{edition}{The edition of the map being described.}

\item{geographicCoverage}{Description of the geographic area which the map covers}

\item{scale}{The Map's scale}

\item{inheritParams}{node_template}
}
\value{
a map list object
}
\description{
map
}
