% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurementScale.R
\name{measurementScale}
\alias{measurementScale}
\title{measurementScale}
\usage{
measurementScale(nominal = NULL, ordinal = NULL, interval = NULL,
  ratio = NULL, dateTime = NULL)
}
\arguments{
\item{nominal}{Characteristics used to define nominal (categorical) scale attributes}

\item{ordinal}{Characteristics used to define ordinal (ordered) scale attributes}

\item{interval}{Characteristics used to define interval scale attributes}

\item{ratio}{Characteristics used to define ratio scale attributes}

\item{dateTime}{Characteristics used to define date and time attributes}

\item{inheritParams}{node_template}
}
\value{
a measurementScale list object
}
\description{
measurementScale
}
