% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\title{methods}
\usage{
methods(methodStep = NULL, sampling = NULL, qualityControl = NULL)
}
\arguments{
\item{methodStep}{Information about the methods employed in collecting or generating a data set or other resource.}

\item{sampling}{Description of sampling procedures including the geographic, temporal and taxonomic coverage of the study.}

\item{qualityControl}{Information on possible errors or on the quality of a data set.}

\item{inheritParams}{node_template}
}
\value{
a methods list object
}
\description{
methods
}
