% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum.R
\name{minimum}
\alias{minimum}
\title{minimum}
\usage{
minimum(exclusive = NULL)
}
\arguments{
\item{exclusive}{Exclusive bounds flag}

\item{inheritParams}{node_template}
}
\value{
a minimum list object
}
\description{
minimum
}
