% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValueCode.R
\name{missingValueCode}
\alias{missingValueCode}
\title{missingValueCode}
\usage{
missingValueCode(code = NULL, codeExplanation = NULL)
}
\arguments{
\item{code}{The missing value code itself.}

\item{codeExplanation}{An explanation of what the missing value code means.}

\item{inheritParams}{node_template}
}
\value{
a missingValueCode list object
}
\description{
missingValueCode
}
