% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDocs.R
\name{moduleDocs}
\alias{moduleDocs}
\title{moduleDocs}
\usage{
moduleDocs(moduleName = NULL, moduleDescription = NULL,
  recommendedUsage = NULL, standAlone = NULL)
}
\arguments{
\item{moduleName}{(def not found)}

\item{moduleDescription}{(def not found)}

\item{recommendedUsage}{(def not found)}

\item{standAlone}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a moduleDocs list object
}
\description{
moduleDocs
}
