% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBand.R
\name{multiBand}
\alias{multiBand}
\title{multiBand}
\usage{
multiBand(nbands = NULL, layout = NULL)
}
\arguments{
\item{nbands}{The number of spectral bands in the image.}

\item{layout}{The organization of the bands in the image file.}

\item{inheritParams}{node_template}
}
\value{
a multiBand list object
}
\description{
multiBand
}
