% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nominal.R
\name{nominal}
\alias{nominal}
\title{nominal}
\usage{
nominal(nonNumericDomain = NULL)
}
\arguments{
\item{nonNumericDomain}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a nominal list object
}
\description{
nominal
}
