% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offline.R
\name{offline}
\alias{offline}
\title{offline}
\usage{
offline(mediumName = NULL, mediumDensity = NULL,
  mediumDensityUnits = NULL, mediumVolume = NULL,
  mediumFormat = NULL, mediumNote = NULL)
}
\arguments{
\item{mediumName}{Name of the medium that for this resource distribution}

\item{mediumDensity}{the density of the digital medium if this is relevant.}

\item{mediumDensityUnits}{a numerical density's units}

\item{mediumVolume}{total volume of the storage medium}

\item{mediumFormat}{format of the medium on which the resource is shipped.}

\item{mediumNote}{note about the media}

\item{inheritParams}{node_template}
}
\value{
a offline list object
}
\description{
offline
}
