% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online.R
\name{online}
\alias{online}
\title{online}
\usage{
online(onlineDescription = NULL, url = NULL, connection = NULL,
  connectionDefinition = NULL)
}
\arguments{
\item{onlineDescription}{onlineDescription}

\item{url}{url}

\item{connection}{connection}

\item{connectionDefinition}{The definition of a connection that will be used in another location in the EML document}

\item{inheritParams}{node_template}

\item{onlineDescription}{Brief description of the the content of online}

\item{url}{A URL or this resource.}

\item{connection}{A connection to a data service}
}
\value{
a online list object
}
\description{
online
}
