% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{ordinal}
\alias{ordinal}
\title{ordinal}
\usage{
ordinal(nonNumericDomain = NULL)
}
\arguments{
\item{nonNumericDomain}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a ordinal list object
}
\description{
ordinal
}
