% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{para}
\alias{para}
\title{para}
\usage{
para(value = NULL, itemizedlist = NULL, orderedlist = NULL,
  emphasis = NULL, subscript = NULL, superscript = NULL,
  literalLayout = NULL, ulink = NULL, lang = NULL)
}
\arguments{
\item{value}{Language translation for the paragraph}

\item{itemizedlist}{A list of items in a text paragraph.}

\item{orderedlist}{An ordered list of items in a text paragraph.}

\item{emphasis}{A span of emphasized text in a paragraph.}

\item{subscript}{A subscript in a text paragraph.}

\item{superscript}{A superscript in a text paragraph.}

\item{literalLayout}{The equivalent to <pre> in html.}

\item{ulink}{a link that addresses its target by means of a URL}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a para list object
}
\description{
para
}
