% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameter}
\alias{parameter}
\title{parameter}
\usage{
parameter(name = NULL, value = NULL, description = NULL,
  domainDescription = NULL, required = NULL, repeats = NULL)
}
\arguments{
\item{name}{The name of the parameter.See \code{\link[=name]{name()}}}

\item{value}{The value supplied for the parameterSee \code{\link[=value]{value()}}}

\item{description}{A description of the Parameter.See \code{\link[=description]{description()}}}

\item{domainDescription}{A description of domain of valid values for the parameter.See \code{\link[=domainDescription]{domainDescription()}}}

\item{required}{Whether or not the parameter is required.See \code{\link[=required]{required()}}}

\item{repeats}{Whether or not the parameter repeats.See \code{\link[=repeats]{repeats()}}}

\item{name}{The name of the parameter.See \code{\link[=name]{name()}}}
}
\value{
a parameter list object
}
\description{
parameter
}
