% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalCommunication.R
\name{personalCommunication}
\alias{personalCommunication}
\title{personalCommunication}
\usage{
personalCommunication(publisher = NULL, publicationPlace = NULL,
  communicationType = NULL, recipient = NULL)
}
\arguments{
\item{publisher}{Organization that actually publishes the communication}

\item{publicationPlace}{The location at which the work was published.}

\item{communicationType}{The type of personal communication.}

\item{recipient}{Recipient of the communication}

\item{inheritParams}{node_template}
}
\value{
a personalCommunication list object
}
\description{
personalCommunication
}
