% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical.R
\name{physical}
\alias{physical}
\title{physical}
\usage{
physical(objectName = NULL, size = NULL, authentication = NULL,
  compressionMethod = NULL, encodingMethod = NULL,
  characterEncoding = NULL, dataFormat = NULL, distribution = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{objectName}{The name of the data object.}

\item{size}{Describes the physical size of the data object.}

\item{authentication}{A value, typically a checksum, used to authenticate that the bitstream delivered to the user is identical to the original.}

\item{compressionMethod}{Name of a compression method applied}

\item{encodingMethod}{Name of a encoding method applied}

\item{characterEncoding}{Contains the name of the character encoding used for the data.}

\item{dataFormat}{Describes the internal physical format of a data object.}

\item{distribution}{Information on how the resource is distributed online and offline}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a physical list object
}
\description{
physical
}
