% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionName.R
\name{positionName}
\alias{positionName}
\title{positionName}
\usage{
positionName(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a positionName list object
}
\description{
positionName
}
