% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postalCode.R
\name{postalCode}
\alias{postalCode}
\title{postalCode}
\usage{
postalCode(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a postalCode list object
}
\description{
postalCode
}
