% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{presentation}
\alias{presentation}
\title{presentation}
\usage{
presentation(conferenceName = NULL, conferenceDate = NULL,
  conferenceLocation = NULL)
}
\arguments{
\item{conferenceName}{The name of the conference at which this presentation was given.}

\item{conferenceDate}{The date the conference was held.}

\item{conferenceLocation}{The location where the conference was held.}

\item{inheritParams}{node_template}
}
\value{
a presentation list object
}
\description{
presentation
}
