% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryKey.R
\name{primaryKey}
\alias{primaryKey}
\title{primaryKey}
\usage{
primaryKey(constraintName = NULL, constraintDescription = NULL,
  key = NULL)
}
\arguments{
\item{constraintName}{A meaningfull name of the constraint.}

\item{constraintDescription}{Descibes the purpose of the constraint.}

\item{key}{The set of attributes to which this constraint applies.}

\item{inheritParams}{node_template}
}
\value{
a primaryKey list object
}
\description{
primaryKey
}
