% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeMeridian.R
\name{primeMeridian}
\alias{primeMeridian}
\title{primeMeridian}
\usage{
primeMeridian(name = NULL, longitude = NULL)
}
\arguments{
\item{name}{The name of the Prime MeridianSee \code{\link[=name]{name()}}}

\item{longitude}{The longitude of the Prime MeridianSee \code{\link[=longitude]{longitude()}}}
}
\value{
a primeMeridian list object
}
\description{
primeMeridian
}
