% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proceduralStep.R
\name{proceduralStep}
\alias{proceduralStep}
\title{proceduralStep}
\usage{
proceduralStep(description = NULL, citation = NULL, protocol = NULL,
  instrumentation = NULL, software = NULL, subStep = NULL)
}
\arguments{
\item{description}{Description of the methods employed in collecting or generating a data set or other resource or in quality control and assurance.}

\item{citation}{Literature citation relating to the methods used.}

\item{protocol}{Protocol description relating to the methods used.}

\item{instrumentation}{Instruments used for measurement and recording data.}

\item{software}{Software used in the processing of data.}

\item{subStep}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a proceduralStep list object
}
\description{
proceduralStep
}
