% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projCoordSys.R
\name{projCoordSys}
\alias{projCoordSys}
\title{projCoordSys}
\usage{
projCoordSys(geogCoordSys = NULL, projection = NULL)
}
\arguments{
\item{geogCoordSys}{The coodinate system used to express the geographic coordinates.}

\item{projection}{The method used to tranform between geographic and planar coordinates.}

\item{inheritParams}{node_template}
}
\value{
a projCoordSys list object
}
\description{
projCoordSys
}
