% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\title{projection}
\usage{
projection(parameter = NULL, unit = NULL, name = NULL)
}
\arguments{
\item{parameter}{A parameter required by the projection method.}

\item{unit}{The unit of measure coordinates are expressed in.}

\item{name}{The name of the projection method.}

\item{inheritParams}{node_template}
}
\value{
a projection list object
}
\description{
projection
}
