% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionList.R
\name{projectionList}
\alias{projectionList}
\title{projectionList}
\usage{
projectionList(horizCoordSysDef = NULL)
}
\arguments{
\item{horizCoordSysDef}{The definition of a coordinate systemSee \code{\link[=horizCoordSysDef]{horizCoordSysDef()}}}
}
\value{
a projectionList list object
}
\description{
projectionList
}
