% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol.R
\name{protocol}
\alias{protocol}
\title{protocol}
\usage{
protocol(alternateIdentifier = NULL, shortName = NULL, title = NULL,
  creator = NULL, metadataProvider = NULL, associatedParty = NULL,
  pubDate = NULL, language = NULL, series = NULL, abstract = NULL,
  keywordSet = NULL, additionalInfo = NULL,
  intellectualRights = NULL, licensed = NULL, distribution = NULL,
  coverage = NULL, annotation = NULL, proceduralStep = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity}

\item{shortName}{A short name that describes the resource, sometimes a filename.}

\item{title}{A brief description of the resource, providing enough detail to differentiate it from other similar resources.}

\item{creator}{The people or organizations who created this resource.}

\item{metadataProvider}{The people or organizations who created provided documentation and other metadata for this resource.}

\item{associatedParty}{Other people or organizations who should be associated with this resource.}

\item{pubDate}{The publication date of the resource.}

\item{language}{The language in which the resource is written.}

\item{series}{The series from which the resource came.}

\item{abstract}{A brief overview of the resource.}

\item{keywordSet}{Keyword information that describes the resource.}

\item{additionalInfo}{Any extra information pertitent to the resource.}

\item{intellectualRights}{Intellectual property rights regarding usage and licensing of this resource.}

\item{licensed}{Information identifying a well-known license for the metadata and data}

\item{distribution}{Information on how the resource is distributed online and offline}

\item{coverage}{Extent of the coverage of the resource.}

\item{annotation}{A precisely-defined semantic statement about this resource.}

\item{proceduralStep}{(def not found)}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a protocol list object
}
\description{
protocol
}
