% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purpose.R
\name{purpose}
\alias{purpose}
\title{purpose}
\usage{
purpose(section = NULL, para = NULL, markdown = NULL, lang = NULL)
}
\arguments{
\item{section}{A section of related text.}

\item{para}{A simple paragraph of text.}

\item{markdown}{A block of text formatted with Markdown directives.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a purpose list object
}
\description{
purpose
}
