% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativeAccuracyReport.R
\name{quantitativeAccuracyReport}
\alias{quantitativeAccuracyReport}
\title{quantitativeAccuracyReport}
\usage{
quantitativeAccuracyReport(quantitativeAccuracyValue = NULL,
  quantitativeAccuracyMethod = NULL)
}
\arguments{
\item{quantitativeAccuracyValue}{The value resulting from the accuracy test.}

\item{quantitativeAccuracyMethod}{The method used to calculate the accuracy value.}

\item{inheritParams}{node_template}
}
\value{
a quantitativeAccuracyReport list object
}
\description{
quantitativeAccuracyReport
}
