% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativeAttributeAccuracyAssessment.R
\name{quantitativeAttributeAccuracyAssessment}
\alias{quantitativeAttributeAccuracyAssessment}
\title{quantitativeAttributeAccuracyAssessment}
\usage{
quantitativeAttributeAccuracyAssessment(attributeAccuracyValue = NULL,
  attributeAccuracyExplanation = NULL)
}
\arguments{
\item{attributeAccuracyValue}{A value assigned to estimate the accuracy of the attribute.}

\item{attributeAccuracyExplanation}{The test which yields the Attribute Accuracy Value.}

\item{inheritParams}{node_template}
}
\value{
a quantitativeAttributeAccuracyAssessment list object
}
\description{
quantitativeAttributeAccuracyAssessment
}
