% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeOfDates.R
\name{rangeOfDates}
\alias{rangeOfDates}
\title{rangeOfDates}
\usage{
rangeOfDates(beginDate = NULL, endDate = NULL)
}
\arguments{
\item{beginDate}{A single time stamp signifying the beginning of some time period}

\item{endDate}{A single time stamp signifying the end of some time period}

\item{inheritParams}{node_template}
}
\value{
a rangeOfDates list object
}
\description{
rangeOfDates
}
