% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio.R
\name{ratio}
\alias{ratio}
\title{ratio}
\usage{
ratio(unit = NULL, precision = NULL, numericDomain = NULL)
}
\arguments{
\item{unit}{(def not found)}

\item{precision}{(def not found)}

\item{numericDomain}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a ratio list object
}
\description{
ratio
}
