% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedKey.R
\name{referencedKey}
\alias{referencedKey}
\title{referencedKey}
\usage{
referencedKey(attributeReference = NULL)
}
\arguments{
\item{attributeReference}{The identifier of an attribute found in the identified entity See \code{\link[=attributeReference]{attributeReference()}}}
}
\value{
a referencedKey list object
}
\description{
referencedKey
}
