% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(reportNumber = NULL, publisher = NULL,
  publicationPlace = NULL, totalPages = NULL)
}
\arguments{
\item{reportNumber}{The unique identification number assigned to the report.}

\item{publisher}{Organization that actually publishes the report}

\item{publicationPlace}{The location at which the work was published.}

\item{totalPages}{The total number of pages in the report.}

\item{inheritParams}{node_template}
}
\value{
a report list object
}
\description{
report
}
