% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{repository}
\alias{repository}
\title{repository}
\usage{
repository(originator = NULL)
}
\arguments{
\item{originator}{A person or organization asociated with this resource.See \code{\link[=originator]{originator()}}}
}
\value{
a repository list object
}
\description{
repository
}
