% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchProject.R
\name{researchProject}
\alias{researchProject}
\title{researchProject}
\usage{
researchProject(title = NULL, personnel = NULL, abstract = NULL,
  funding = NULL, award = NULL, studyAreaDescription = NULL,
  designDescription = NULL, relatedProject = NULL, references = NULL,
  id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{title}{Title of the project.}

\item{personnel}{Contact and role information for people involved in the research project.}

\item{abstract}{Project Abstract.}

\item{funding}{Funding information.}

\item{award}{Award information.}

\item{studyAreaDescription}{Description of the physical area associated with the research project, potentially including coverage, climate, geology, disturbances, etc.}

\item{designDescription}{Description of the design of the research project}

\item{relatedProject}{This field is a recursive link to another project.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a researchProject list object
}
\description{
researchProject
}
