% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salutation.R
\name{salutation}
\alias{salutation}
\title{salutation}
\usage{
salutation(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a salutation list object
}
\description{
salutation
}
