% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingDescription.R
\name{samplingDescription}
\alias{samplingDescription}
\title{samplingDescription}
\usage{
samplingDescription(section = NULL, para = NULL, markdown = NULL,
  lang = NULL)
}
\arguments{
\item{section}{A section of related text.}

\item{para}{A simple paragraph of text.}

\item{markdown}{A block of text formatted with Markdown directives.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a samplingDescription list object
}
\description{
samplingDescription
}
