% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section}
\alias{section}
\title{section}
\usage{
section(title = NULL, para = NULL, section = NULL, lang = NULL)
}
\arguments{
\item{title}{The optional title of the section. See \code{\link[=title]{title()}}}

\item{para}{A simple paragraph of text. See \code{\link[=para]{para()}}}

\item{section}{A section of related text. See \code{\link[=section]{section()}}}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a section list object
}
\description{
section
}
