% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDelimited.R
\name{simpleDelimited}
\alias{simpleDelimited}
\title{simpleDelimited}
\usage{
simpleDelimited(fieldDelimiter = NULL, collapseDelimiters = NULL,
  quoteCharacter = NULL, literalCharacter = NULL)
}
\arguments{
\item{fieldDelimiter}{Character used to delimit the end of an attribute}

\item{collapseDelimiters}{Specification of how to handle consecutive delimiters while parsing}

\item{quoteCharacter}{Character used to quote values for delimiter escaping}

\item{literalCharacter}{Character used to escape other special characters}

\item{inheritParams}{node_template}
}
\value{
a simpleDelimited list object
}
\description{
simpleDelimited
}
