% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{size}
\usage{
size(unit = NULL)
}
\arguments{
\item{unit}{Unit of measurement for the entity size, by default byteSee \code{\link[=unit]{unit()}}}
}
\value{
a size list object
}
\description{
size
}
