% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialRaster.R
\name{spatialRaster}
\alias{spatialRaster}
\title{spatialRaster}
\usage{
spatialRaster(alternateIdentifier = NULL, entityName = NULL,
  entityDescription = NULL, physical = NULL, coverage = NULL,
  methods = NULL, additionalInfo = NULL, annotation = NULL,
  attributeList = NULL, constraint = NULL, spatialReference = NULL,
  georeferenceInfo = NULL, horizontalAccuracy = NULL,
  verticalAccuracy = NULL, cellSizeXDirection = NULL,
  cellSizeYDirection = NULL, numberOfBands = NULL,
  rasterOrigin = NULL, rows = NULL, columns = NULL,
  verticals = NULL, cellGeometry = NULL, toneGradation = NULL,
  scaleFactor = NULL, offset = NULL, imageDescription = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity}

\item{entityName}{The name of the entity name}

\item{entityDescription}{General description of the entity and its contents}

\item{physical}{Description of the physical format of the entity}

\item{coverage}{Information on the geographic, spatial and temporal coverages used in this entity.}

\item{methods}{(def not found)}

\item{additionalInfo}{Any extra information pertitent to the entity.}

\item{annotation}{A precisely-defined semantic statement about this entity.}

\item{attributeList}{The list of attributes associated with this entity.}

\item{constraint}{Description of any relational constraints on this entity.}

\item{spatialReference}{The means by which positional information in this entity is related to points on the earth's surface.}

\item{georeferenceInfo}{Information on how to position the grid within the coordinate system defined in the spatial reference module.}

\item{horizontalAccuracy}{The accuracy of horizontal locational measurements within the data.}

\item{verticalAccuracy}{The accuracy of vertical locational measurements within the data.}

\item{cellSizeXDirection}{The width of the cell in the x direction.}

\item{cellSizeYDirection}{The width of the cell in the x direction.}

\item{numberOfBands}{The number of bands in the image.}

\item{rasterOrigin}{The corner location of the pixel having the minimum x and y values}

\item{rows}{Maximum number of raster objects along the ordinate (y) axis}

\item{columns}{Maximum number of raster objects along the abscissa (x) axis}

\item{verticals}{Maximum number of raster objects along the vertical (z) axis}

\item{cellGeometry}{Geometric representation of the cell's content.}

\item{toneGradation}{Number of colors present in the image.}

\item{scaleFactor}{The value used for scaling the source raster.}

\item{offset}{The offset value used for scaling the source raster.}

\item{imageDescription}{Detailed information for data derived from image sensors.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a spatialRaster list object
}
\description{
spatialRaster
}
