% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialReference.R
\name{spatialReference}
\alias{spatialReference}
\title{spatialReference}
\usage{
spatialReference(horizCoordSysName = NULL, horizCoordSysDef = NULL,
  vertCoordSys = NULL, references = NULL, id = NULL, system = NULL,
  scope = NULL)
}
\arguments{
\item{horizCoordSysName}{The name of a predefined coordinate system}

\item{horizCoordSysDef}{The definition of a coordinate system}

\item{vertCoordSys}{The reference frame or system from which vertical distances (altitudes or depths) are measured.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a spatialReference list object
}
\description{
spatialReference
}
