% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialVector.R
\name{spatialVector}
\alias{spatialVector}
\title{spatialVector}
\usage{
spatialVector(alternateIdentifier = NULL, entityName = NULL,
  entityDescription = NULL, physical = NULL, coverage = NULL,
  methods = NULL, additionalInfo = NULL, annotation = NULL,
  attributeList = NULL, constraint = NULL, geometry = NULL,
  geometricObjectCount = NULL, topologyLevel = NULL,
  spatialReference = NULL, horizontalAccuracy = NULL,
  verticalAccuracy = NULL, references = NULL, id = NULL,
  system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity}

\item{entityName}{The name of the entity name}

\item{entityDescription}{General description of the entity and its contents}

\item{physical}{Description of the physical format of the entity}

\item{coverage}{Information on the geographic, spatial and temporal coverages used in this entity.}

\item{methods}{(def not found)}

\item{additionalInfo}{Any extra information pertitent to the entity.}

\item{annotation}{A precisely-defined semantic statement about this entity.}

\item{attributeList}{The list of attributes associated with this entity.}

\item{constraint}{Description of any relational constraints on this entity.}

\item{geometry}{Type of geometric feature(s) represented in the entity.}

\item{geometricObjectCount}{Total number of the geometric objects occurring in the dataset.}

\item{topologyLevel}{Code which identifies the degree of complexity.}

\item{spatialReference}{(def not found)}

\item{horizontalAccuracy}{Horizontal positional accuracy of the data}

\item{verticalAccuracy}{Vertical positional accuracy of the data}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a spatialVector list object
}
\description{
spatialVector
}
