% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{spheroid}
\alias{spheroid}
\title{spheroid}
\usage{
spheroid(name = NULL, semiAxisMajor = NULL, denomFlatRatio = NULL)
}
\arguments{
\item{name}{The name of the spheroid.}

\item{semiAxisMajor}{The larger of two orthogonal axes that define the ellipse}

\item{denomFlatRatio}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a spheroid list object
}
\description{
spheroid
}
