% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storageType.R
\name{storageType}
\alias{storageType}
\title{storageType}
\usage{
storageType(typeSystem = NULL)
}
\arguments{
\item{typeSystem}{The system used to define the storage types. This should be an identifier of a well known and published typing system.See \code{\link[=typeSystem]{typeSystem()}}}
}
\value{
a storageType list object
}
\description{
storageType
}
