% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedProcedure.R
\name{storedProcedure}
\alias{storedProcedure}
\title{storedProcedure}
\usage{
storedProcedure(alternateIdentifier = NULL, entityName = NULL,
  entityDescription = NULL, physical = NULL, coverage = NULL,
  methods = NULL, additionalInfo = NULL, annotation = NULL,
  attributeList = NULL, constraint = NULL, parameter = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity}

\item{entityName}{The name of the entity name}

\item{entityDescription}{General description of the entity and its contents}

\item{physical}{Description of the physical format of the entity}

\item{coverage}{Information on the geographic, spatial and temporal coverages used in this entity.}

\item{methods}{(def not found)}

\item{additionalInfo}{Any extra information pertitent to the entity.}

\item{annotation}{A precisely-defined semantic statement about this entity.}

\item{attributeList}{The list of attributes associated with this entity.}

\item{constraint}{(def not found)}

\item{parameter}{A parameter that is required as input to the stored procedure.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a storedProcedure list object
}
\description{
storedProcedure
}
