% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyExtent.R
\name{studyExtent}
\alias{studyExtent}
\title{studyExtent}
\usage{
studyExtent(coverage = NULL, description = NULL)
}
\arguments{
\item{coverage}{A description of the geographic area sampled (geographic coverage), the sampling frequency (temporal coverage), and living organisms sampled (taxonomic coverage.}

\item{description}{A textual description of the specific areas sampled (geographic coverage), the sampling frequency (temporal coverage), and groups of living organisms sampled (taxonomic coverage).}

\item{inheritParams}{node_template}
}
\value{
a studyExtent list object
}
\description{
studyExtent
}
