% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscript.R
\name{subscript}
\alias{subscript}
\title{subscript}
\usage{
subscript(value = NULL, subscript = NULL, superscript = NULL,
  lang = NULL)
}
\arguments{
\item{value}{Language translation for the sub/superscript}

\item{subscript}{A subscript in a text paragraph.}

\item{superscript}{A superscript in a text paragraph.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a subscript list object
}
\description{
subscript
}
