% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superscript.R
\name{superscript}
\alias{superscript}
\title{superscript}
\usage{
superscript(value = NULL, subscript = NULL, superscript = NULL,
  lang = NULL)
}
\arguments{
\item{value}{Language translation for the sub/superscript}

\item{subscript}{A subscript in a text paragraph.}

\item{superscript}{A superscript in a text paragraph.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a superscript list object
}
\description{
superscript
}
