% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomicCoverage.R
\name{taxonomicCoverage}
\alias{taxonomicCoverage}
\title{taxonomicCoverage}
\usage{
taxonomicCoverage(taxonomicSystem = NULL,
  generalTaxonomicCoverage = NULL, taxonomicClassification = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{taxonomicSystem}{Documentation of taxonomic sources, procedures, and treatments.}

\item{generalTaxonomicCoverage}{A description of the range of taxa addressed in the data set or collection.}

\item{taxonomicClassification}{Information about the range of taxa addressed in the data set or collection.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a taxonomicCoverage list object
}
\description{
taxonomicCoverage
}
