% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalCoverage.R
\name{temporalCoverage}
\alias{temporalCoverage}
\title{temporalCoverage}
\usage{
temporalCoverage(singleDateTime = NULL, rangeOfDates = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{singleDateTime}{Means of encoding a single date and time}

\item{rangeOfDates}{Means of encoding a range of dates and times.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a temporalCoverage list object
}
\description{
temporalCoverage
}
