% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFormat.R
\name{textFormat}
\alias{textFormat}
\title{textFormat}
\usage{
textFormat(numHeaderLines = NULL, numFooterLines = NULL,
  recordDelimiter = NULL, physicalLineDelimiter = NULL,
  numPhysicalLinesPerRecord = NULL, maxRecordLength = NULL,
  attributeOrientation = NULL, simpleDelimited = NULL,
  complex = NULL)
}
\arguments{
\item{numHeaderLines}{Number of header lines preceding data.}

\item{numFooterLines}{Number of footer lines following data.}

\item{recordDelimiter}{Character used to delimit records.}

\item{physicalLineDelimiter}{Character used to delimit physical lines.}

\item{numPhysicalLinesPerRecord}{The number of physical lines in the file spanned by a single logical data record.}

\item{maxRecordLength}{The maximum number of characters in any record in the physical file.}

\item{attributeOrientation}{Orientation of attributes.}

\item{simpleDelimited}{A simple delimited format.}

\item{complex}{A complex text format.}

\item{inheritParams}{node_template}
}
\value{
a textFormat list object
}
\description{
textFormat
}
