% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis}
\alias{thesis}
\title{thesis}
\usage{
thesis(degree = NULL, institution = NULL, totalPages = NULL)
}
\arguments{
\item{degree}{The name or degree level for which the thesis was completed. See \code{\link[=degree]{degree()}}}

\item{institution}{The name of the degree-awarding institution See \code{\link[=institution]{institution()}}}

\item{totalPages}{The total number of pages in the thesis. See \code{\link[=totalPages]{totalPages()}}}
}
\value{
a thesis list object
}
\description{
thesis
}
