% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{unit}
\alias{unit}
\title{unit}
\usage{
unit(standardUnit = NULL, customUnit = NULL, name = NULL)
}
\arguments{
\item{standardUnit}{The name of a standard unit used to make this measurementSee \code{\link[=standardUnit]{standardUnit()}}}

\item{customUnit}{The name of a custom unit used to make this measurement.See \code{\link[=customUnit]{customUnit()}}}

\item{name}{The name of the unit of measure.See \code{\link[=name]{name()}}}
}
\value{
a unit list object
}
\description{
unit
}
