% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usageCitation.R
\name{usageCitation}
\alias{usageCitation}
\title{usageCitation}
\usage{
usageCitation(alternateIdentifier = NULL, shortName = NULL,
  title = NULL, creator = NULL, metadataProvider = NULL,
  associatedParty = NULL, pubDate = NULL, language = NULL,
  series = NULL, abstract = NULL, keywordSet = NULL,
  additionalInfo = NULL, intellectualRights = NULL, licensed = NULL,
  distribution = NULL, coverage = NULL, annotation = NULL,
  contact = NULL, article = NULL, book = NULL, chapter = NULL,
  editedBook = NULL, manuscript = NULL, report = NULL,
  thesis = NULL, conferenceProceedings = NULL,
  personalCommunication = NULL, map = NULL, generic = NULL,
  audioVisual = NULL, presentation = NULL, bibtex = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity}

\item{shortName}{A short name that describes the resource, sometimes a filename.}

\item{title}{A brief description of the resource, providing enough detail to differentiate it from other similar resources.}

\item{creator}{The people or organizations who created this resource.}

\item{metadataProvider}{The people or organizations who created provided documentation and other metadata for this resource.}

\item{associatedParty}{Other people or organizations who should be associated with this resource.}

\item{pubDate}{The publication date of the resource.}

\item{language}{The language in which the resource is written.}

\item{series}{The series from which the resource came.}

\item{abstract}{A brief overview of the resource.}

\item{keywordSet}{Keyword information that describes the resource.}

\item{additionalInfo}{Any extra information pertitent to the resource.}

\item{intellectualRights}{Intellectual property rights regarding usage and licensing of this resource.}

\item{licensed}{Information identifying a well-known license for the metadata and data}

\item{distribution}{Information on how the resource is distributed online and offline}

\item{coverage}{Extent of the coverage of the resource.}

\item{annotation}{A precisely-defined semantic statement about this resource.}

\item{contact}{An optional contact individual for this citation}

\item{article}{Information for full citation of a journal article or other periodical article.}

\item{book}{Information for full citation of a book}

\item{chapter}{Information for citation of a chapter in a book}

\item{editedBook}{Information for full citation of an edited book}

\item{manuscript}{Information about an unpublished manuscript}

\item{report}{Information about a report published by an institution.}

\item{thesis}{Information about a thesis that has been written as part of a degree requirement and is frequently published in small numbers by the degree awarding institution.}

\item{conferenceProceedings}{The published notes, papers, presentations etc... of a conference.}

\item{personalCommunication}{A personal communication that has been saved online or as hard-copy.}

\item{map}{This element describes the map that is being cited or cataloged.}

\item{generic}{This reference type was created for references that do not fit in to the other existing reference types}

\item{audioVisual}{This reference type is meant to cover all forms of audio and visual media.}

\item{presentation}{An unpublished presentation from a conference, workshop, workgroup, symposium etc.}

\item{bibtex}{Citation in Bibtex format.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a usageCitation list object
}
\description{
usageCitation
}
