% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userId.R
\name{userId}
\alias{userId}
\title{userId}
\usage{
userId(directory = NULL)
}
\arguments{
\item{directory}{The directory system within which the directoryId can be retrieved.}

\item{inheritParams}{node_template}
}
\value{
a userId list object
}
\description{
userId
}
