% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{value}
\usage{
value(lang = NULL)
}
\arguments{
\item{lang}{(def not found)}

\item{inheritParams}{node_template}

\item{lang}{language code for element content}
}
\value{
a value list object
}
\description{
value
}
