% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertCoordSys.R
\name{vertCoordSys}
\alias{vertCoordSys}
\title{vertCoordSys}
\usage{
vertCoordSys(altitudeSysDef = NULL, depthSysDef = NULL)
}
\arguments{
\item{altitudeSysDef}{the reference frame or system from which altitudes (elevations) are measured. See \code{\link[=altitudeSysDef]{altitudeSysDef()}}}

\item{depthSysDef}{The reference frame or system from which depths are measured. See \code{\link[=depthSysDef]{depthSysDef()}}}
}
\value{
a vertCoordSys list object
}
\description{
vertCoordSys
}
