% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vouchers.R
\name{vouchers}
\alias{vouchers}
\title{vouchers}
\usage{
vouchers(specimen = NULL, repository = NULL)
}
\arguments{
\item{specimen}{(def not found)}

\item{repository}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a vouchers list object
}
\description{
vouchers
}
