% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access}
\alias{access}
\title{access}
\usage{
access(allow = NULL, deny = NULL, references = NULL, id = NULL,
  system = NULL, scope = NULL, order = NULL, authSystem = NULL)
}
\arguments{
\item{allow}{A rule that grants a permission type.}

\item{deny}{A rule that revokes a permission type.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{order}{The order in which the allow and deny rules should be applied.}

\item{authSystem}{The authentication system is used to verify the user or group to whom access is allowed or denied.}

\item{inheritParams}{node_template}
}
\value{
a access list object
}
\description{
access
}
