% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalInfo.R
\name{additionalInfo}
\alias{additionalInfo}
\title{additionalInfo}
\usage{
additionalInfo(section = NULL, para = NULL, markdown = NULL,
  lang = NULL)
}
\arguments{
\item{section}{A section of related text.}

\item{para}{A simple paragraph of text.}

\item{markdown}{A block of text formatted with Markdown directives.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a additionalInfo list object
}
\description{
additionalInfo
}
