% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrativeArea.R
\name{administrativeArea}
\alias{administrativeArea}
\title{administrativeArea}
\usage{
administrativeArea(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a administrativeArea list object
}
\description{
administrativeArea
}
