% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternateIdentifier.R
\name{alternateIdentifier}
\alias{alternateIdentifier}
\title{alternateIdentifier}
\usage{
alternateIdentifier(system = NULL)
}
\arguments{
\item{inheritParams}{node_template}
}
\value{
a alternateIdentifier list object
}
\description{
alternateIdentifier
}
