% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotation}
\alias{annotation}
\title{annotation}
\usage{
annotation(propertyURI = NULL, propertyLabel = NULL, valueURI = NULL,
  valueLabel = NULL)
}
\arguments{
\item{propertyURI}{The persistent URI used to identify a property from a vocabulary.}

\item{propertyLabel}{A human readable representation of the controlled property.}

\item{valueURI}{The persistent URI used to identify a value from a vocabulary.}

\item{valueLabel}{A human readable representation of the controlled value.}

\item{inheritParams}{node_template}
}
\value{
a annotation list object
}
\description{
annotation
}
