% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{article}
\alias{article}
\title{article}
\usage{
article(journal = NULL, volume = NULL, issue = NULL,
  pageRange = NULL, publisher = NULL, publicationPlace = NULL,
  ISSN = NULL)
}
\arguments{
\item{journal}{The name of the journal, magazine, newspaper, zine, etc... in which the article was published.See \code{\link[=journal]{journal()}}}

\item{volume}{The volume of the journal in which the article appearsSee \code{\link[=volume]{volume()}}}

\item{issue}{The issue of the journal in which the article appearsSee \code{\link[=issue]{issue()}}}

\item{pageRange}{The beginning and end page numbers of the journal articleSee \code{\link[=pageRange]{pageRange()}}}

\item{publisher}{Organization that actually publishes the articleSee \code{\link[=publisher]{publisher()}}}

\item{publicationPlace}{The location at which the work was published.See \code{\link[=publicationPlace]{publicationPlace()}}}

\item{ISSN}{The unique Internation Standard Serial NumberSee \code{\link[=ISSN]{ISSN()}}}
}
\value{
a article list object
}
\description{
article
}
