% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{attribute}
\alias{attribute}
\title{attribute}
\usage{
attribute(attributeName = NULL, attributeLabel = NULL,
  attributeDefinition = NULL, storageType = NULL,
  measurementScale = NULL, missingValueCode = NULL, accuracy = NULL,
  coverage = NULL, methods = NULL, annotation = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{attributeName}{The name of the attribute}

\item{attributeLabel}{A label for displaying an attribute name.}

\item{attributeDefinition}{Precise definition of the attribute}

\item{storageType}{Storage type hint for this field}

\item{measurementScale}{The measurement scale for the attribute.}

\item{missingValueCode}{Character for missing value in the data of the field}

\item{accuracy}{The accuracy of the attribute. This information should describe any accuracy information that is known about the collection of this data attribute.}

\item{coverage}{An explanation of the coverage of the attribute.}

\item{methods}{An explanation of the methods involved in the collection of this attribute.}

\item{annotation}{A precisely-defined semantic statement for this attribute.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a attribute list object
}
\description{
attribute
}
