% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioVisual.R
\name{audioVisual}
\alias{audioVisual}
\title{audioVisual}
\usage{
audioVisual(publisher = NULL, publicationPlace = NULL,
  performer = NULL, ISBN = NULL)
}
\arguments{
\item{publisher}{Organization which actually distributes the video, film, the broadcaster etc...}

\item{publicationPlace}{The location at which the work was published.}

\item{performer}{The performers in the audio visual production}

\item{ISBN}{The unique Internation Standard Book Number}

\item{inheritParams}{node_template}
}
\value{
a audioVisual list object
}
\description{
audioVisual
}
