% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/award.R
\name{award}
\alias{award}
\title{award}
\usage{
award(funderName = NULL, funderIdentifier = NULL, awardNumber = NULL,
  title = NULL, awardUrl = NULL)
}
\arguments{
\item{funderName}{The name of the funding institution that made this award}

\item{funderIdentifier}{The identifier of the funding agency.}

\item{awardNumber}{The assigned award number.}

\item{title}{The title of the award.}

\item{awardUrl}{The URL associated with award.}

\item{inheritParams}{node_template}
}
\value{
a award list object
}
\description{
award
}
