% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awardUrl.R
\name{awardUrl}
\alias{awardUrl}
\title{awardUrl}
\usage{
awardUrl(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a awardUrl list object
}
\description{
awardUrl
}
