% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandDescription.R
\name{bandDescription}
\alias{bandDescription}
\title{bandDescription}
\usage{
bandDescription(sequenceIdentifier = NULL, highWavelength = NULL,
  lowWaveLength = NULL, waveLengthUnits = NULL, peakResponse = NULL)
}
\arguments{
\item{sequenceIdentifier}{Number that uniquely identifies instances of bands of wavelengths on which a sensor operates.See \code{\link[=sequenceIdentifier]{sequenceIdentifier()}}}

\item{highWavelength}{Highest wavelength that the sensor is capable of collecting within a designated band in metres.See \code{\link[=highWavelength]{highWavelength()}}}

\item{lowWaveLength}{Lowest wavelength that the sensor is capable of collecting within a designated band in metres.See \code{\link[=lowWaveLength]{lowWaveLength()}}}

\item{waveLengthUnits}{Units in which the sensor wavelegths are expressed.See \code{\link[=waveLengthUnits]{waveLengthUnits()}}}

\item{peakResponse}{Wavelength at which the response is the highest.See \code{\link[=peakResponse]{peakResponse()}}}
}
\value{
a bandDescription list object
}
\description{
bandDescription
}
