% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundingAltitudes.R
\name{boundingAltitudes}
\alias{boundingAltitudes}
\title{boundingAltitudes}
\usage{
boundingAltitudes(altitudeMinimum = NULL, altitudeMaximum = NULL,
  altitudeUnits = NULL)
}
\arguments{
\item{altitudeMinimum}{The minimum altitude extent of coverage. See \code{\link[=altitudeMinimum]{altitudeMinimum()}}}

\item{altitudeMaximum}{The maximum altitude extent of coverage. See \code{\link[=altitudeMaximum]{altitudeMaximum()}}}

\item{altitudeUnits}{The unit of altitude See \code{\link[=altitudeUnits]{altitudeUnits()}}}
}
\value{
a boundingAltitudes list object
}
\description{
boundingAltitudes
}
