% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds.R
\name{bounds}
\alias{bounds}
\title{bounds}
\usage{
bounds(minimum = NULL, maximum = NULL)
}
\arguments{
\item{minimum}{Minimum numeric bound of attributeSee \code{\link[=minimum]{minimum()}}}

\item{maximum}{Maximum numeric bound of attributeSee \code{\link[=maximum]{maximum()}}}
}
\value{
a bounds list object
}
\description{
bounds
}
