% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeHistory.R
\name{changeHistory}
\alias{changeHistory}
\title{changeHistory}
\usage{
changeHistory(changeScope = NULL, oldValue = NULL, changeDate = NULL,
  comment = NULL)
}
\arguments{
\item{changeScope}{An expression describing the scope to which the documented change was applied.See \code{\link[=changeScope]{changeScope()}}}

\item{oldValue}{The previous value of the data prior to the change.See \code{\link[=oldValue]{oldValue()}}}

\item{changeDate}{The date the changes were applied.See \code{\link[=changeDate]{changeDate()}}}

\item{comment}{Explanation or justification for the change made to the data.See \code{\link[=comment]{comment()}}}
}
\value{
a changeHistory list object
}
\description{
changeHistory
}
