% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationSystem.R
\name{classificationSystem}
\alias{classificationSystem}
\title{classificationSystem}
\usage{
classificationSystem(classificationSystemCitation = NULL,
  classificationSystemModifications = NULL)
}
\value{
a classificationSystem list object
}
\description{
classificationSystem
}
