% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeDefinition.R
\name{codeDefinition}
\alias{codeDefinition}
\title{codeDefinition}
\usage{
codeDefinition(code = NULL, definition = NULL, source = NULL,
  order = NULL)
}
\arguments{
\item{code}{Code value allowed in the domain}

\item{definition}{Definition of the associated code}

\item{source}{The name of the source for this code and its definition}

\item{order}{Mechanism for specifying what the order of the code-definitions included should be}

\item{inheritParams}{node_template}
}
\value{
a codeDefinition list object
}
\description{
codeDefinition
}
