% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferenceLocation.R
\name{conferenceLocation}
\alias{conferenceLocation}
\title{conferenceLocation}
\usage{
conferenceLocation(deliveryPoint = NULL, city = NULL,
  administrativeArea = NULL, postalCode = NULL, country = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{deliveryPoint}{The location for postal deliveries.}

\item{city}{The name of the city for the contact.}

\item{administrativeArea}{The political area of a country.}

\item{postalCode}{The postal code used for routing to an address.}

\item{country}{The name of the country for the contact's address.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a conferenceLocation list object
}
\description{
conferenceLocation
}
