% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conferenceProceedings.R
\name{conferenceProceedings}
\alias{conferenceProceedings}
\title{conferenceProceedings}
\usage{
conferenceProceedings(publisher = NULL, publicationPlace = NULL,
  edition = NULL, volume = NULL, numberOfVolumes = NULL,
  totalPages = NULL, totalFigures = NULL, totalTables = NULL,
  ISBN = NULL, chapterNumber = NULL, editor = NULL,
  bookTitle = NULL, pageRange = NULL, conferenceName = NULL,
  conferenceDate = NULL, conferenceLocation = NULL)
}
\arguments{
\item{publisher}{Organization that actually publishes the book}

\item{publicationPlace}{The location at which the work was published.}

\item{edition}{The edition of the book being described.}

\item{volume}{The volume of the book that is part of a series.}

\item{numberOfVolumes}{Number of volumes in a collection}

\item{totalPages}{The total number of pages in the book.}

\item{totalFigures}{The total number of figures in the book.}

\item{totalTables}{The total number of tables in a book.}

\item{ISBN}{The unique Internation Standard Book Number}

\item{chapterNumber}{The chapter number of interest within a book}

\item{editor}{The name of the editor of the book.}

\item{bookTitle}{The title of the book.}

\item{pageRange}{The beginning and ending page numbers of a chapter.}

\item{conferenceName}{The name of the conference whose proceedings have been published.}

\item{conferenceDate}{The date the conference was held.}

\item{conferenceLocation}{The location where the conference was held.}

\item{inheritParams}{node_template}
}
\value{
a conferenceProceedings list object
}
\description{
conferenceProceedings
}
