% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionDefinition.R
\name{connectionDefinition}
\alias{connectionDefinition}
\title{connectionDefinition}
\usage{
connectionDefinition(schemeName = NULL, description = NULL,
  parameterDefinition = NULL, references = NULL, id = NULL,
  system = NULL, scope = NULL)
}
\arguments{
\item{schemeName}{The name of the scheme used to identify this connection.}

\item{description}{The description of the scheme used to identify this connection.}

\item{parameterDefinition}{The definition of a parameter that is needed to properly use this connection scheme.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a connectionDefinition list object
}
\description{
connectionDefinition
}
