% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.R
\name{constraint}
\alias{constraint}
\title{constraint}
\usage{
constraint(primaryKey = NULL, uniqueKey = NULL,
  checkConstraint = NULL, foreignKey = NULL, joinCondition = NULL,
  notNullConstraint = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{primaryKey}{The primary key in the entity See \code{\link[=primaryKey]{primaryKey()}}}

\item{uniqueKey}{A unique key in the entity See \code{\link[=uniqueKey]{uniqueKey()}}}

\item{checkConstraint}{A constraint which checks a conditional clause within an entity. See \code{\link[=checkConstraint]{checkConstraint()}}}

\item{foreignKey}{A foreign key relationship among entities See \code{\link[=foreignKey]{foreignKey()}}}

\item{joinCondition}{A non primary/foreign key join See \code{\link[=joinCondition]{joinCondition()}}}

\item{notNullConstraint}{A constraint that indicates that no null values should be present for an attribute. See \code{\link[=notNullConstraint]{notNullConstraint()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}
}
\value{
a constraint list object
}
\description{
constraint
}
