% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornerPoint.R
\name{cornerPoint}
\alias{cornerPoint}
\title{cornerPoint}
\usage{
cornerPoint(xCoordinate = NULL, yCoordinate = NULL,
  pointInPixel = NULL, corner = NULL)
}
\arguments{
\item{xCoordinate}{X Location of the georeferencing point in the X diminsion of the coordinate system.}

\item{yCoordinate}{Location of the georeferencing point in the Y dimension of the coordinate system.}

\item{pointInPixel}{Location within the pixel of the georeferencing point.}

\item{corner}{Identification of the corner in the grid corresponding to the coordinates provided.}

\item{inheritParams}{node_template}
}
\value{
a cornerPoint list object
}
\description{
cornerPoint
}
