% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{country}
\alias{country}
\title{country}
\usage{
country(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a country list object
}
\description{
country
}
