% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormat.R
\name{dataFormat}
\alias{dataFormat}
\title{dataFormat}
\usage{
dataFormat(textFormat = NULL, externallyDefinedFormat = NULL,
  binaryRasterFormat = NULL)
}
\arguments{
\item{textFormat}{Description of a text formatted objectSee \code{\link[=textFormat]{textFormat()}}}

\item{externallyDefinedFormat}{Information about a non-text or proprietary formatted object.See \code{\link[=externallyDefinedFormat]{externallyDefinedFormat()}}}

\item{binaryRasterFormat}{Contains binary raster data header parametersSee \code{\link[=binaryRasterFormat]{binaryRasterFormat()}}}
}
\value{
a dataFormat list object
}
\description{
dataFormat
}
