% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTime.R
\name{dateTime}
\alias{dateTime}
\title{dateTime}
\usage{
dateTime(formatString = NULL, dateTimePrecision = NULL,
  dateTimeDomain = NULL)
}
\arguments{
\item{formatString}{A format string that describes the format for a date-time value from the Gregorian calendar.}

\item{dateTimePrecision}{An indication of the precision of a date or time value}

\item{dateTimeDomain}{See the summary for the type: DateTimeDomainType}

\item{inheritParams}{node_template}
}
\value{
a dateTime list object
}
\description{
dateTime
}
