% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeDomain.R
\name{dateTimeDomain}
\alias{dateTimeDomain}
\title{dateTimeDomain}
\usage{
dateTimeDomain(bounds = NULL, references = NULL, id = NULL)
}
\arguments{
\item{bounds}{Elements for specifying allowed date range.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a dateTimeDomain list object
}
\description{
dateTimeDomain
}
