% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveryPoint.R
\name{deliveryPoint}
\alias{deliveryPoint}
\title{deliveryPoint}
\usage{
deliveryPoint(value = NULL, lang = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a deliveryPoint list object
}
\description{
deliveryPoint
}
