% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depthSysDef.R
\name{depthSysDef}
\alias{depthSysDef}
\title{depthSysDef}
\usage{
depthSysDef(depthDatumName = NULL, depthResolution = NULL,
  depthDistanceUnits = NULL, depthEncodingMethod = NULL)
}
\arguments{
\item{depthDatumName}{The identification given to surface of reference from which depths are measured. See \code{\link[=depthDatumName]{depthDatumName()}}}

\item{depthResolution}{The minimum distance possible between two adjacent depth values, expressed in Depth Distance Units of measure. See \code{\link[=depthResolution]{depthResolution()}}}

\item{depthDistanceUnits}{The units in which depths are recorded. See \code{\link[=depthDistanceUnits]{depthDistanceUnits()}}}

\item{depthEncodingMethod}{The means used to encode depths. See \code{\link[=depthEncodingMethod]{depthEncodingMethod()}}}
}
\value{
a depthSysDef list object
}
\description{
depthSysDef
}
