% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electronicMailAddress.R
\name{electronicMailAddress}
\alias{electronicMailAddress}
\title{electronicMailAddress}
\usage{
electronicMailAddress(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a electronicMailAddress list object
}
\description{
electronicMailAddress
}
