% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityCodeList.R
\name{entityCodeList}
\alias{entityCodeList}
\title{entityCodeList}
\usage{
entityCodeList(entityReference = NULL, valueAttributeReference = NULL,
  definitionAttributeReference = NULL, orderAttributeReference = NULL)
}
\arguments{
\item{entityReference}{A reference to the id of the entity in which the code list has been defined See \code{\link[=entityReference]{entityReference()}}}

\item{valueAttributeReference}{A reference to the id of the attribute that contains the list of codes See \code{\link[=valueAttributeReference]{valueAttributeReference()}}}

\item{definitionAttributeReference}{A reference to the id of the attribute that contains the definition of codes See \code{\link[=definitionAttributeReference]{definitionAttributeReference()}}}

\item{orderAttributeReference}{A reference to the id of the attribute that contains the order of codes See \code{\link[=orderAttributeReference]{orderAttributeReference()}}}
}
\value{
a entityCodeList list object
}
\description{
entityCodeList
}
