% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityCodeList.R
\name{entityCodeList}
\alias{entityCodeList}
\title{entityCodeList}
\usage{
entityCodeList(entityReference = NULL, valueAttributeReference = NULL,
  definitionAttributeReference = NULL, orderAttributeReference = NULL)
}
\arguments{
\item{entityReference}{A reference to the id of the entity in which the code list has been defined}

\item{valueAttributeReference}{A reference to the id of the attribute that contains the list of codes}

\item{definitionAttributeReference}{A reference to the id of the attribute that contains the definition of codes}

\item{orderAttributeReference}{A reference to the id of the attribute that contains the order of codes}

\item{inheritParams}{node_template}
}
\value{
a entityCodeList list object
}
\description{
entityCodeList
}
