% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeratedDomain.R
\name{enumeratedDomain}
\alias{enumeratedDomain}
\title{enumeratedDomain}
\usage{
enumeratedDomain(codeDefinition = NULL, externalCodeSet = NULL,
  entityCodeList = NULL, enforced = NULL)
}
\arguments{
\item{codeDefinition}{A codes and its definitionSee \code{\link[=codeDefinition]{codeDefinition()}}}

\item{externalCodeSet}{A reference to an externally defined set of codes used in this attributeSee \code{\link[=externalCodeSet]{externalCodeSet()}}}

\item{entityCodeList}{A code list that is defined in a data tableSee \code{\link[=entityCodeList]{entityCodeList()}}}

\item{enforced}{Indicates whether the enumerated domain values enforced.See \code{\link[=enforced]{enforced()}}}
}
\value{
a enumeratedDomain list object
}
\description{
enumeratedDomain
}
