% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example}
\alias{example}
\title{example}
\usage{
example(section = NULL, para = NULL, markdown = NULL, lang = NULL)
}
\arguments{
\item{section}{A section of related text. See \code{\link[=section]{section()}}}

\item{para}{A simple paragraph of text. See \code{\link[=para]{para()}}}

\item{markdown}{A block of text formatted with Markdown directives. See \code{\link[=markdown]{markdown()}}}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a example list object
}
\description{
example
}
