% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCodeSet.R
\name{externalCodeSet}
\alias{externalCodeSet}
\title{externalCodeSet}
\usage{
externalCodeSet(codesetName = NULL, citation = NULL,
  codesetURL = NULL)
}
\arguments{
\item{codesetName}{The name of an externally defined code setSee \code{\link[=codesetName]{codesetName()}}}

\item{citation}{A citation for the code set referenceSee \code{\link[=citation]{citation()}}}

\item{codesetURL}{A URL for the code set referenceSee \code{\link[=codesetURL]{codesetURL()}}}
}
\value{
a externalCodeSet list object
}
\description{
externalCodeSet
}
