% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externalCodeSet.R
\name{externalCodeSet}
\alias{externalCodeSet}
\title{externalCodeSet}
\usage{
externalCodeSet(codesetName = NULL, citation = NULL,
  codesetURL = NULL)
}
\arguments{
\item{codesetName}{The name of an externally defined code set}

\item{citation}{A citation for the code set reference}

\item{codesetURL}{A URL for the code set reference}

\item{inheritParams}{node_template}
}
\value{
a externalCodeSet list object
}
\description{
externalCodeSet
}
