% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/externallyDefinedFormat.R
\name{externallyDefinedFormat}
\alias{externallyDefinedFormat}
\title{externallyDefinedFormat}
\usage{
externallyDefinedFormat(formatName = NULL, formatVersion = NULL,
  citation = NULL)
}
\arguments{
\item{formatName}{Name of the format of the data object See \code{\link[=formatName]{formatName()}}}

\item{formatVersion}{Version of the format of the data object See \code{\link[=formatVersion]{formatVersion()}}}

\item{citation}{Citation providing more details about the physical format. See \code{\link[=citation]{citation()}}}
}
\value{
a externallyDefinedFormat list object
}
\description{
externallyDefinedFormat
}
