% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funderIdentifier.R
\name{funderIdentifier}
\alias{funderIdentifier}
\title{funderIdentifier}
\usage{
funderIdentifier(value = NULL, lang = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a funderIdentifier list object
}
\description{
funderIdentifier
}
