% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funderName.R
\name{funderName}
\alias{funderName}
\title{funderName}
\usage{
funderName(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a funderName list object
}
\description{
funderName
}
