% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{generic}
\alias{generic}
\title{generic}
\usage{
generic(publisher = NULL, publicationPlace = NULL,
  referenceType = NULL, volume = NULL, numberOfVolumes = NULL,
  totalPages = NULL, totalFigures = NULL, totalTables = NULL,
  edition = NULL, originalPublication = NULL, reprintEdition = NULL,
  reviewedItem = NULL, ISBN = NULL, ISSN = NULL)
}
\arguments{
\item{publisher}{Organization which actually publishes the reference}

\item{publicationPlace}{The location at which the work was published.}

\item{referenceType}{The type of reference.}

\item{volume}{The volume of the reference that is part of a series.}

\item{numberOfVolumes}{Number of volumes in a collection}

\item{totalPages}{The total number of pages in the references.}

\item{totalFigures}{The total number of figures in the reference.}

\item{totalTables}{The total number of tables in a reference.}

\item{edition}{The edition of the generic reference being described.}

\item{originalPublication}{References current publication to its original.}

\item{reprintEdition}{Reference for current edition that was originally published under a different title.}

\item{reviewedItem}{Reference types that are reviews of other references.}

\item{ISBN}{The unique Internation Standard Book Number}

\item{ISSN}{The unique Internation Standard Serial Number}

\item{inheritParams}{node_template}
}
\value{
a generic list object
}
\description{
generic
}
