% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geogCoordSys.R
\name{geogCoordSys}
\alias{geogCoordSys}
\title{geogCoordSys}
\usage{
geogCoordSys(datum = NULL, spheroid = NULL, primeMeridian = NULL,
  unit = NULL, name = NULL)
}
\arguments{
\item{datum}{The point on the earth representing the origin for the geographic coordinates to which the coorinate system is related to.See \code{\link[=datum]{datum()}}}

\item{spheroid}{An ellipse that is used to model the earth's surface.See \code{\link[=spheroid]{spheroid()}}}

\item{unit}{The unit of angle by which coordinates are expressed.See \code{\link[=unit]{unit()}}}
}
\value{
a geogCoordSys list object
}
\description{
geogCoordSys
}
