% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRingPoint.R
\name{gRingPoint}
\alias{gRingPoint}
\title{gRingPoint}
\usage{
gRingPoint(gRingLatitude = NULL, gRingLongitude = NULL)
}
\arguments{
\item{gRingLatitude}{The latitude of a point of the g-ring.}

\item{gRingLongitude}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a gRingPoint list object
}
\description{
gRingPoint
}
