% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementation.R
\name{implementation}
\alias{implementation}
\title{implementation}
\usage{
implementation(distribution = NULL, size = NULL, language = NULL,
  operatingSystem = NULL, machineProcessor = NULL,
  virtualMachine = NULL, diskUsage = NULL, runtimeMemoryUsage = NULL,
  programmingLanguage = NULL, checksum = NULL, dependency = NULL)
}
\arguments{
\item{distribution}{Information on how the resource is distributed online and offline}

\item{size}{Physical size of an implementation.}

\item{language}{The International Language of the software implementation.}

\item{operatingSystem}{The operating system(s) an implementation runs on.}

\item{machineProcessor}{The machine processor(s) required for executing the implementation.}

\item{virtualMachine}{The virtual machine that the implementation requires.}

\item{diskUsage}{The minimum amount of Disk Space required to install this implementation.}

\item{runtimeMemoryUsage}{The minimum amount of memory required to run an implementation.}

\item{programmingLanguage}{The computer programming language the software package was programmed in.}

\item{checksum}{The generated checksum value of a software package that is a self-contained module.}

\item{dependency}{This fields documents any dependencies that this implementation might have.}

\item{inheritParams}{node_template}
}
\value{
a implementation list object
}
\description{
implementation
}
