% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword.R
\name{keyword}
\alias{keyword}
\title{keyword}
\usage{
keyword(value = NULL, lang = NULL, keywordType = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}

\item{keywordType}{The type of each keyword.}
}
\value{
a keyword list object
}
\description{
keyword
}
