% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywordSet.R
\name{keywordSet}
\alias{keywordSet}
\title{keywordSet}
\usage{
keywordSet(keyword = NULL, keywordThesaurus = NULL)
}
\arguments{
\item{keyword}{A single keyword that describes the resource. See \code{\link[=keyword]{keyword()}}}

\item{keywordThesaurus}{The name of a thesaurus from which the keyword is derived. See \code{\link[=keywordThesaurus]{keywordThesaurus()}}}
}
\value{
a keywordSet list object
}
\description{
keywordSet
}
