% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensed.R
\name{licensed}
\alias{licensed}
\title{licensed}
\usage{
licensed(licenseName = NULL, url = NULL, identifier = NULL)
}
\arguments{
\item{licenseName}{The official name of the license See \code{\link[=licenseName]{licenseName()}}}

\item{url}{The persistent URL for the license See \code{\link[=url]{url()}}}

\item{identifier}{License Identifier See \code{\link[=identifier]{identifier()}}}
}
\value{
a licensed list object
}
\description{
licensed
}
