% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listitem.R
\name{listitem}
\alias{listitem}
\title{listitem}
\usage{
listitem(para = NULL, itemizedlist = NULL, orderedlist = NULL)
}
\arguments{
\item{para}{A simple paragraph of text.}

\item{itemizedlist}{A list of items in a text paragraph.}

\item{orderedlist}{An ordered list of items in a text paragraph.}

\item{inheritParams}{node_template}
}
\value{
a listitem list object
}
\description{
listitem
}
