% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literalLayout.R
\name{literalLayout}
\alias{literalLayout}
\title{literalLayout}
\usage{
literalLayout(value = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a literalLayout list object
}
\description{
literalLayout
}
