% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literatureCited.R
\name{literatureCited}
\alias{literatureCited}
\title{literatureCited}
\usage{
literatureCited(citation = NULL, bibtex = NULL)
}
\arguments{
\item{citation}{A citation within a list of citations. See \code{\link[=citation]{citation()}}}

\item{bibtex}{List of citations in Bibtex format. See \code{\link[=bibtex]{bibtex()}}}
}
\value{
a literatureCited list object
}
\description{
literatureCited
}
