% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintenance.R
\name{maintenance}
\alias{maintenance}
\title{maintenance}
\usage{
maintenance(description = NULL, maintenanceUpdateFrequency = NULL,
  changeHistory = NULL)
}
\arguments{
\item{description}{A text description of the maintenance of this data resource.}

\item{maintenanceUpdateFrequency}{Frequency with which changes and additions are made to the dataset after the initial dataset is completed.}

\item{changeHistory}{A description of changes made to the data since its release.}

\item{inheritParams}{node_template}
}
\value{
a maintenance list object
}
\description{
maintenance
}
