% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript.R
\name{manuscript}
\alias{manuscript}
\title{manuscript}
\usage{
manuscript(institution = NULL, totalPages = NULL)
}
\arguments{
\item{institution}{The address and contact information needed to request a manuscript.}

\item{totalPages}{The total number of pages in the manuscript.}

\item{inheritParams}{node_template}
}
\value{
a manuscript list object
}
\description{
manuscript
}
