% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{map}
\usage{
map(publisher = NULL, edition = NULL, geographicCoverage = NULL,
  scale = NULL)
}
\arguments{
\item{publisher}{Organization that actually publishes the map See \code{\link[=publisher]{publisher()}}}

\item{edition}{The edition of the map being described. See \code{\link[=edition]{edition()}}}

\item{geographicCoverage}{Description of the geographic area which the map covers See \code{\link[=geographicCoverage]{geographicCoverage()}}}

\item{scale}{The Map's scale See \code{\link[=scale]{scale()}}}
}
\value{
a map list object
}
\description{
map
}
