% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\title{methods}
\usage{
methods(methodStep = NULL, sampling = NULL, qualityControl = NULL)
}
\arguments{
\item{methodStep}{Information about the methods employed in collecting or generating a data set or other resource.See \code{\link[=methodStep]{methodStep()}}}

\item{sampling}{Description of sampling procedures including the geographic, temporal and taxonomic coverage of the study.See \code{\link[=sampling]{sampling()}}}

\item{qualityControl}{Information on possible errors or on the quality of a data set.See \code{\link[=qualityControl]{qualityControl()}}}
}
\value{
a methods list object
}
\description{
methods
}
