% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValueCode.R
\name{missingValueCode}
\alias{missingValueCode}
\title{missingValueCode}
\usage{
missingValueCode(code = NULL, codeExplanation = NULL)
}
\arguments{
\item{code}{The missing value code itself. See \code{\link[=code]{code()}}}

\item{codeExplanation}{An explanation of what the missing value code means. See \code{\link[=codeExplanation]{codeExplanation()}}}
}
\value{
a missingValueCode list object
}
\description{
missingValueCode
}
