% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDocs.R
\name{moduleDocs}
\alias{moduleDocs}
\title{moduleDocs}
\usage{
moduleDocs(moduleName = NULL, moduleDescription = NULL,
  recommendedUsage = NULL, standAlone = NULL)
}
\value{
a moduleDocs list object
}
\description{
moduleDocs
}
