% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBand.R
\name{multiBand}
\alias{multiBand}
\title{multiBand}
\usage{
multiBand(nbands = NULL, layout = NULL)
}
\arguments{
\item{nbands}{The number of spectral bands in the image.See \code{\link[=nbands]{nbands()}}}

\item{layout}{The organization of the bands in the image file.See \code{\link[=layout]{layout()}}}
}
\value{
a multiBand list object
}
\description{
multiBand
}
