% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonNumericDomain.R
\name{nonNumericDomain}
\alias{nonNumericDomain}
\title{nonNumericDomain}
\usage{
nonNumericDomain(enumeratedDomain = NULL, textDomain = NULL,
  references = NULL, id = NULL)
}
\arguments{
\item{enumeratedDomain}{Description of any coded values associated with the attribute.}

\item{textDomain}{Description of a free-text domain pattern for the attribute}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a nonNumericDomain list object
}
\description{
nonNumericDomain
}
