% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericDomain.R
\name{numericDomain}
\alias{numericDomain}
\title{numericDomain}
\usage{
numericDomain(numberType = NULL, bounds = NULL, references = NULL,
  id = NULL)
}
\arguments{
\item{numberType}{(def not found)}

\item{bounds}{Elements for specifying allowed values range.}

\item{references}{The id of another element in this EML document to be used to here in this context.}

\item{inheritParams}{node_template}
}
\value{
a numericDomain list object
}
\description{
numericDomain
}
