% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online.R
\name{online}
\alias{online}
\title{online}
\usage{
online(onlineDescription = NULL, url = NULL, connection = NULL,
  connectionDefinition = NULL)
}
\arguments{
\item{onlineDescription}{onlineDescriptionSee \code{\link[=onlineDescription]{onlineDescription()}}}

\item{url}{urlSee \code{\link[=url]{url()}}}

\item{connection}{connectionSee \code{\link[=connection]{connection()}}}

\item{connectionDefinition}{The definition of a connection that will be used in another location in the EML documentSee \code{\link[=connectionDefinition]{connectionDefinition()}}}
}
\value{
a online list object
}
\description{
online
}
