% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineDescription.R
\name{onlineDescription}
\alias{onlineDescription}
\title{onlineDescription}
\usage{
onlineDescription(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a onlineDescription list object
}
\description{
onlineDescription
}
