% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderedlist.R
\name{orderedlist}
\alias{orderedlist}
\title{orderedlist}
\usage{
orderedlist(listitem = NULL)
}
\arguments{
\item{listitem}{An item in a list of items. See \code{\link[=listitem]{listitem()}}}
}
\value{
a orderedlist list object
}
\description{
orderedlist
}
