% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizationName.R
\name{organizationName}
\alias{organizationName}
\title{organizationName}
\usage{
organizationName(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a organizationName list object
}
\description{
organizationName
}
