% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para.R
\name{para}
\alias{para}
\title{para}
\usage{
para(value = NULL, itemizedlist = NULL, orderedlist = NULL,
  emphasis = NULL, subscript = NULL, superscript = NULL,
  literalLayout = NULL, ulink = NULL, lang = NULL)
}
\arguments{
\item{value}{Language translation for the paragraphSee \code{\link[=value]{value()}}}

\item{itemizedlist}{A list of items in a text paragraph.See \code{\link[=itemizedlist]{itemizedlist()}}}

\item{orderedlist}{An ordered list of items in a text paragraph.See \code{\link[=orderedlist]{orderedlist()}}}

\item{emphasis}{A span of emphasized text in a paragraph.See \code{\link[=emphasis]{emphasis()}}}

\item{subscript}{A subscript in a text paragraph.See \code{\link[=subscript]{subscript()}}}

\item{superscript}{A superscript in a text paragraph.See \code{\link[=superscript]{superscript()}}}

\item{literalLayout}{The equivalent to <pre> in html.See \code{\link[=literalLayout]{literalLayout()}}}

\item{ulink}{a link that addresses its target by means of a URLSee \code{\link[=ulink]{ulink()}}}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a para list object
}
\description{
para
}
