% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{parameter}
\alias{parameter}
\title{parameter}
\usage{
parameter(name = NULL, value = NULL, description = NULL,
  domainDescription = NULL, required = NULL, repeats = NULL)
}
\arguments{
\item{name}{Name of the parameter to be used to make this connection.}

\item{value}{The value of the parameter to be used to make this connection.}

\item{description}{A description of the Parameter.}

\item{domainDescription}{A description of domain of valid values for the parameter.}

\item{required}{Whether or not the parameter is required.}

\item{repeats}{Whether or not the parameter repeats.}

\item{inheritParams}{node_template}

\item{name}{The name of the parameter.}

\item{value}{The value supplied for the parameter}

\item{name}{The name of the parameter.}
}
\value{
a parameter list object
}
\description{
parameter
}
