% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterDefinition.R
\name{parameterDefinition}
\alias{parameterDefinition}
\title{parameterDefinition}
\usage{
parameterDefinition(name = NULL, definition = NULL,
  defaultValue = NULL)
}
\arguments{
\item{name}{The Name of a parameter that is needed to properly use this connection scheme.}

\item{definition}{The definition of a parameter that is needed to properly use this connection scheme.}

\item{defaultValue}{The default value for a parameter that is needed to properly use this connection scheme.}

\item{inheritParams}{node_template}
}
\value{
a parameterDefinition list object
}
\description{
parameterDefinition
}
