% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalCommunication.R
\name{personalCommunication}
\alias{personalCommunication}
\title{personalCommunication}
\usage{
personalCommunication(publisher = NULL, publicationPlace = NULL,
  communicationType = NULL, recipient = NULL)
}
\arguments{
\item{publisher}{Organization that actually publishes the communicationSee \code{\link[=publisher]{publisher()}}}

\item{publicationPlace}{The location at which the work was published.See \code{\link[=publicationPlace]{publicationPlace()}}}

\item{communicationType}{The type of personal communication.See \code{\link[=communicationType]{communicationType()}}}

\item{recipient}{Recipient of the communicationSee \code{\link[=recipient]{recipient()}}}
}
\value{
a personalCommunication list object
}
\description{
personalCommunication
}
