% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical.R
\name{physical}
\alias{physical}
\title{physical}
\usage{
physical(objectName = NULL, size = NULL, authentication = NULL,
  compressionMethod = NULL, encodingMethod = NULL,
  characterEncoding = NULL, dataFormat = NULL, distribution = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{objectName}{The name of the data object. See \code{\link[=objectName]{objectName()}}}

\item{size}{Describes the physical size of the data object. See \code{\link[=size]{size()}}}

\item{authentication}{A value, typically a checksum, used to authenticate that the bitstream delivered to the user is identical to the original. See \code{\link[=authentication]{authentication()}}}

\item{compressionMethod}{Name of a compression method applied See \code{\link[=compressionMethod]{compressionMethod()}}}

\item{encodingMethod}{Name of a encoding method applied See \code{\link[=encodingMethod]{encodingMethod()}}}

\item{characterEncoding}{Contains the name of the character encoding used for the data. See \code{\link[=characterEncoding]{characterEncoding()}}}

\item{dataFormat}{Describes the internal physical format of a data object. See \code{\link[=dataFormat]{dataFormat()}}}

\item{distribution}{Information on how the resource is distributed online and offline See \code{\link[=distribution]{distribution()}}}

\item{references}{The id of another element in this EML document to be used to here in this context. See \code{\link[=references]{references()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}
}
\value{
a physical list object
}
\description{
physical
}
