% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionName.R
\name{positionName}
\alias{positionName}
\title{positionName}
\usage{
positionName(value = NULL, lang = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a positionName list object
}
\description{
positionName
}
