% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primeMeridian.R
\name{primeMeridian}
\alias{primeMeridian}
\title{primeMeridian}
\usage{
primeMeridian(name = NULL, longitude = NULL)
}
\arguments{
\item{name}{The name of the Prime Meridian}

\item{longitude}{The longitude of the Prime Meridian}

\item{inheritParams}{node_template}
}
\value{
a primeMeridian list object
}
\description{
primeMeridian
}
