% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projCoordSys.R
\name{projCoordSys}
\alias{projCoordSys}
\title{projCoordSys}
\usage{
projCoordSys(geogCoordSys = NULL, projection = NULL)
}
\arguments{
\item{geogCoordSys}{The coodinate system used to express the geographic coordinates.See \code{\link[=geogCoordSys]{geogCoordSys()}}}

\item{projection}{The method used to tranform between geographic and planar coordinates.See \code{\link[=projection]{projection()}}}
}
\value{
a projCoordSys list object
}
\description{
projCoordSys
}
