% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitativeAttributeAccuracyAssessment.R
\name{quantitativeAttributeAccuracyAssessment}
\alias{quantitativeAttributeAccuracyAssessment}
\title{quantitativeAttributeAccuracyAssessment}
\usage{
quantitativeAttributeAccuracyAssessment(attributeAccuracyValue = NULL,
  attributeAccuracyExplanation = NULL)
}
\arguments{
\item{attributeAccuracyValue}{A value assigned to estimate the accuracy of the attribute. See \code{\link[=attributeAccuracyValue]{attributeAccuracyValue()}}}

\item{attributeAccuracyExplanation}{The test which yields the Attribute Accuracy Value. See \code{\link[=attributeAccuracyExplanation]{attributeAccuracyExplanation()}}}
}
\value{
a quantitativeAttributeAccuracyAssessment list object
}
\description{
quantitativeAttributeAccuracyAssessment
}
