% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(reportNumber = NULL, publisher = NULL,
  publicationPlace = NULL, totalPages = NULL)
}
\arguments{
\item{reportNumber}{The unique identification number assigned to the report.See \code{\link[=reportNumber]{reportNumber()}}}

\item{publisher}{Organization that actually publishes the reportSee \code{\link[=publisher]{publisher()}}}

\item{publicationPlace}{The location at which the work was published.See \code{\link[=publicationPlace]{publicationPlace()}}}

\item{totalPages}{The total number of pages in the report.See \code{\link[=totalPages]{totalPages()}}}
}
\value{
a report list object
}
\description{
report
}
