% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salutation.R
\name{salutation}
\alias{salutation}
\title{salutation}
\usage{
salutation(value = NULL, lang = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a salutation list object
}
\description{
salutation
}
