% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section}
\alias{section}
\title{section}
\usage{
section(title = NULL, para = NULL, section = NULL, lang = NULL)
}
\arguments{
\item{title}{The optional title of the section.}

\item{para}{A simple paragraph of text.}

\item{section}{A section of related text.}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a section list object
}
\description{
section
}
