% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleDelimited.R
\name{simpleDelimited}
\alias{simpleDelimited}
\title{simpleDelimited}
\usage{
simpleDelimited(fieldDelimiter = NULL, collapseDelimiters = NULL,
  quoteCharacter = NULL, literalCharacter = NULL)
}
\arguments{
\item{fieldDelimiter}{Character used to delimit the end of an attributeSee \code{\link[=fieldDelimiter]{fieldDelimiter()}}}

\item{collapseDelimiters}{Specification of how to handle consecutive delimiters while parsingSee \code{\link[=collapseDelimiters]{collapseDelimiters()}}}

\item{quoteCharacter}{Character used to quote values for delimiter escapingSee \code{\link[=quoteCharacter]{quoteCharacter()}}}

\item{literalCharacter}{Character used to escape other special charactersSee \code{\link[=literalCharacter]{literalCharacter()}}}
}
\value{
a simpleDelimited list object
}
\description{
simpleDelimited
}
