% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleDateTime.R
\name{singleDateTime}
\alias{singleDateTime}
\title{singleDateTime}
\usage{
singleDateTime(calendarDate = NULL, time = NULL,
  alternativeTimeScale = NULL)
}
\arguments{
\item{calendarDate}{The calendar date for an event.}

\item{time}{The time of day for an event.}

\item{alternativeTimeScale}{A name, code, or date describing an event or period in an alternative time scale, such as one of the geologic time scales.}

\item{inheritParams}{node_template}
}
\value{
a singleDateTime list object
}
\description{
singleDateTime
}
