% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialRaster.R
\name{spatialRaster}
\alias{spatialRaster}
\title{spatialRaster}
\usage{
spatialRaster(alternateIdentifier = NULL, entityName = NULL,
  entityDescription = NULL, physical = NULL, coverage = NULL,
  methods = NULL, additionalInfo = NULL, annotation = NULL,
  attributeList = NULL, constraint = NULL, spatialReference = NULL,
  georeferenceInfo = NULL, horizontalAccuracy = NULL,
  verticalAccuracy = NULL, cellSizeXDirection = NULL,
  cellSizeYDirection = NULL, numberOfBands = NULL,
  rasterOrigin = NULL, rows = NULL, columns = NULL,
  verticals = NULL, cellGeometry = NULL, toneGradation = NULL,
  scaleFactor = NULL, offset = NULL, imageDescription = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity See \code{\link[=alternateIdentifier]{alternateIdentifier()}}}

\item{entityName}{The name of the entity name See \code{\link[=entityName]{entityName()}}}

\item{entityDescription}{General description of the entity and its contents See \code{\link[=entityDescription]{entityDescription()}}}

\item{physical}{Description of the physical format of the entity See \code{\link[=physical]{physical()}}}

\item{coverage}{Information on the geographic, spatial and temporal coverages used in this entity. See \code{\link[=coverage]{coverage()}}}

\item{additionalInfo}{Any extra information pertitent to the entity. See \code{\link[=additionalInfo]{additionalInfo()}}}

\item{annotation}{A precisely-defined semantic statement about this entity. See \code{\link[=annotation]{annotation()}}}

\item{attributeList}{The list of attributes associated with this entity. See \code{\link[=attributeList]{attributeList()}}}

\item{constraint}{Description of any relational constraints on this entity. See \code{\link[=constraint]{constraint()}}}

\item{spatialReference}{The means by which positional information in this entity is related to points on the earth's surface. See \code{\link[=spatialReference]{spatialReference()}}}

\item{georeferenceInfo}{Information on how to position the grid within the coordinate system defined in the spatial reference module. See \code{\link[=georeferenceInfo]{georeferenceInfo()}}}

\item{horizontalAccuracy}{The accuracy of horizontal locational measurements within the data. See \code{\link[=horizontalAccuracy]{horizontalAccuracy()}}}

\item{verticalAccuracy}{The accuracy of vertical locational measurements within the data. See \code{\link[=verticalAccuracy]{verticalAccuracy()}}}

\item{cellSizeXDirection}{The width of the cell in the x direction. See \code{\link[=cellSizeXDirection]{cellSizeXDirection()}}}

\item{cellSizeYDirection}{The width of the cell in the x direction. See \code{\link[=cellSizeYDirection]{cellSizeYDirection()}}}

\item{numberOfBands}{The number of bands in the image. See \code{\link[=numberOfBands]{numberOfBands()}}}

\item{rasterOrigin}{The corner location of the pixel having the minimum x and y values See \code{\link[=rasterOrigin]{rasterOrigin()}}}

\item{rows}{Maximum number of raster objects along the ordinate (y) axis See \code{\link[=rows]{rows()}}}

\item{columns}{Maximum number of raster objects along the abscissa (x) axis See \code{\link[=columns]{columns()}}}

\item{verticals}{Maximum number of raster objects along the vertical (z) axis See \code{\link[=verticals]{verticals()}}}

\item{cellGeometry}{Geometric representation of the cell's content. See \code{\link[=cellGeometry]{cellGeometry()}}}

\item{toneGradation}{Number of colors present in the image. See \code{\link[=toneGradation]{toneGradation()}}}

\item{scaleFactor}{The value used for scaling the source raster. See \code{\link[=scaleFactor]{scaleFactor()}}}

\item{offset}{The offset value used for scaling the source raster. See \code{\link[=offset]{offset()}}}

\item{imageDescription}{Detailed information for data derived from image sensors. See \code{\link[=imageDescription]{imageDescription()}}}

\item{references}{The id of another element in this EML document to be used to here in this context. See \code{\link[=references]{references()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}
}
\value{
a spatialRaster list object
}
\description{
spatialRaster
}
