% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSamplingUnits.R
\name{spatialSamplingUnits}
\alias{spatialSamplingUnits}
\title{spatialSamplingUnits}
\usage{
spatialSamplingUnits(referencedEntityId = NULL, coverage = NULL)
}
\arguments{
\item{referencedEntityId}{The identifier of an entity described in the entity module. This is usually a GIS layer. See \code{\link[=referencedEntityId]{referencedEntityId()}}}

\item{coverage}{Structured description of each sampling unit location See \code{\link[=coverage]{coverage()}}}
}
\value{
a spatialSamplingUnits list object
}
\description{
spatialSamplingUnits
}
