% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedProcedure.R
\name{storedProcedure}
\alias{storedProcedure}
\title{storedProcedure}
\usage{
storedProcedure(alternateIdentifier = NULL, entityName = NULL,
  entityDescription = NULL, physical = NULL, coverage = NULL,
  methods = NULL, additionalInfo = NULL, annotation = NULL,
  attributeList = NULL, constraint = NULL, parameter = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{alternateIdentifier}{A secondary identifier for this entity See \code{\link[=alternateIdentifier]{alternateIdentifier()}}}

\item{entityName}{The name of the entity name See \code{\link[=entityName]{entityName()}}}

\item{entityDescription}{General description of the entity and its contents See \code{\link[=entityDescription]{entityDescription()}}}

\item{physical}{Description of the physical format of the entity See \code{\link[=physical]{physical()}}}

\item{coverage}{Information on the geographic, spatial and temporal coverages used in this entity. See \code{\link[=coverage]{coverage()}}}

\item{additionalInfo}{Any extra information pertitent to the entity. See \code{\link[=additionalInfo]{additionalInfo()}}}

\item{annotation}{A precisely-defined semantic statement about this entity. See \code{\link[=annotation]{annotation()}}}

\item{attributeList}{The list of attributes associated with this entity. See \code{\link[=attributeList]{attributeList()}}}

\item{parameter}{A parameter that is required as input to the stored procedure. See \code{\link[=parameter]{parameter()}}}

\item{references}{The id of another element in this EML document to be used to here in this context. See \code{\link[=references]{references()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}
}
\value{
a storedProcedure list object
}
\description{
storedProcedure
}
