% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studyAreaDescription.R
\name{studyAreaDescription}
\alias{studyAreaDescription}
\title{studyAreaDescription}
\usage{
studyAreaDescription(descriptor = NULL, citation = NULL,
  coverage = NULL)
}
\arguments{
\item{descriptor}{Description of non-coverage characteristics of the study area such as climate, geology, disturbances}

\item{citation}{The citation for this descriptor.}

\item{coverage}{The coverage of this descriptor.}

\item{inheritParams}{node_template}
}
\value{
a studyAreaDescription list object
}
\description{
studyAreaDescription
}
