% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscript.R
\name{subscript}
\alias{subscript}
\title{subscript}
\usage{
subscript(value = NULL, subscript = NULL, superscript = NULL,
  lang = NULL)
}
\arguments{
\item{value}{Language translation for the sub/superscript See \code{\link[=value]{value()}}}

\item{subscript}{A subscript in a text paragraph. See \code{\link[=subscript]{subscript()}}}

\item{superscript}{A superscript in a text paragraph. See \code{\link[=superscript]{superscript()}}}

\item{lang}{Language (an \code{xml:lang} attribute) of the provided text.}
}
\value{
a subscript list object
}
\description{
subscript
}
