% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surName.R
\name{surName}
\alias{surName}
\title{surName}
\usage{
surName(value = NULL, lang = NULL)
}
\arguments{
\item{value}{(def not found)}

\item{lang}{(def not found)}

\item{inheritParams}{node_template}
}
\value{
a surName list object
}
\description{
surName
}
