% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomicCoverage.R
\name{taxonomicCoverage}
\alias{taxonomicCoverage}
\title{taxonomicCoverage}
\usage{
taxonomicCoverage(taxonomicSystem = NULL,
  generalTaxonomicCoverage = NULL, taxonomicClassification = NULL,
  references = NULL, id = NULL, system = NULL, scope = NULL)
}
\arguments{
\item{taxonomicSystem}{Documentation of taxonomic sources, procedures, and treatments. See \code{\link[=taxonomicSystem]{taxonomicSystem()}}}

\item{generalTaxonomicCoverage}{A description of the range of taxa addressed in the data set or collection. See \code{\link[=generalTaxonomicCoverage]{generalTaxonomicCoverage()}}}

\item{taxonomicClassification}{Information about the range of taxa addressed in the data set or collection. See \code{\link[=taxonomicClassification]{taxonomicClassification()}}}

\item{references}{The id of another element in this EML document to be used to here in this context. See \code{\link[=references]{references()}}}

\item{id}{A unique identifier for this additional metadata that can
be used to reference it elsewhere. This is a formal field in that
it is an error to provide a value for the id attribute that is not
unique within the document's set of id attributes. This is designed
to allow other portions of the metadata to reference this section formally.}

\item{system}{The data management system within which an identifier
is in scope and therefore unique. This is typically a URL
(Uniform Resource Locator) that indicates a data management system.
All identifiers that share a system must be unique. In other words,
if the same identifier is used in two locations with identical
systems, then by definition the objects at which they point are
in fact the same object.}

\item{scope}{The scope of the identifier. Scope is generally set
to either "system", meaning that it is scoped according to the
"system" attribute, or "document" if it is only to be in scope
within this single document instance. In this particular use
of scope, it is FIXED to be "system" because the packageId
is required and always has the scope of the required "system".}
}
\value{
a taxonomicCoverage list object
}
\description{
taxonomicCoverage
}
