% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomicSystem.R
\name{taxonomicSystem}
\alias{taxonomicSystem}
\title{taxonomicSystem}
\usage{
taxonomicSystem(classificationSystem = NULL,
  identificationReference = NULL, identifierName = NULL,
  taxonomicProcedures = NULL, taxonomicCompleteness = NULL,
  vouchers = NULL)
}
\arguments{
\item{classificationSystem}{Information about the classification system or authority used.See \code{\link[=classificationSystem]{classificationSystem()}}}
}
\value{
a taxonomicSystem list object
}
\description{
taxonomicSystem
}
