% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textDomain.R
\name{textDomain}
\alias{textDomain}
\title{textDomain}
\usage{
textDomain(definition = NULL, pattern = NULL, source = NULL)
}
\arguments{
\item{definition}{Definition of what this text domain represents}

\item{pattern}{Regular expression pattern constraining the attribute}

\item{source}{The name of the source for this text domain.}

\item{inheritParams}{node_template}
}
\value{
a textDomain list object
}
\description{
textDomain
}
