% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title}
\alias{title}
\title{title}
\usage{
title(value = NULL, lang = NULL)
}
\arguments{
\item{value}{The text value (e.g. when given text is devined for
multiple languages.)}

\item{lang}{language code for element contentSee \code{\link[=lang]{lang()}}}
}
\value{
a title list object
}
\description{
title
}
