% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulink.R
\name{ulink}
\alias{ulink}
\title{ulink}
\usage{
ulink(citetitle = NULL, url = NULL)
}
\arguments{
\item{citetitle}{the title of the cited workSee \code{\link[=citetitle]{citetitle()}}}

\item{url}{the Uniform Resource Locator for the cited workSee \code{\link[=url]{url()}}}
}
\value{
a ulink list object
}
\description{
ulink
}
