% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{value}
\usage{
value(lang = NULL)
}
\arguments{
\item{lang}{language code for element contentSee \code{\link[=lang]{lang()}}}
}
\value{
a value list object
}
\description{
value
}
