% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticalAccuracy.R
\name{verticalAccuracy}
\alias{verticalAccuracy}
\title{verticalAccuracy}
\usage{
verticalAccuracy(accuracyReport = NULL,
  quantitativeAccuracyReport = NULL)
}
\arguments{
\item{accuracyReport}{A qualitative statement about the accuracy of the data.}

\item{quantitativeAccuracyReport}{A quantitative assessment of the data quality.}

\item{inheritParams}{node_template}
}
\value{
a verticalAccuracy list object
}
\description{
verticalAccuracy
}
